% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adding_a_Text_Box_at_the_Top_of_a_Histogram.R
\name{Adding_a_Text_Box_at_the_Top_of_a_Histogram}
\alias{Adding_a_Text_Box_at_the_Top_of_a_Histogram}
\title{Adding a Text Box at the Top of a Histogram}
\usage{
Adding_a_Text_Box_at_the_Top_of_a_Histogram(
  Values,
  Text,
  Text_Size = 1,
  Plot_Margins = c(5, 4, 4, 2) + 0.1,
  Tolerance = sqrt(.Machine$double.eps),
  ...
)
}
\arguments{
\item{Values}{the values to use to generate the histogram.}

\item{Text}{the text to be added to the histogram. This argument may be of class \code{'character'}, \code{'language'}, \code{'expression'}, or \code{'call'}.}

\item{Text_Size}{the size of the desired text in the text box. The default for this argument is \code{1}.}

\item{Plot_Margins}{the plot margins, which are passed to the \code{mar} argument of the \code{par()} function. The default value for this argument is the default plot margins, \code{c(5, 4, 4, 2) + 0.1}.}

\item{Tolerance}{how close together vertical-axis limits from consecutive iterations must be for the algorithm to converge to a solution. The default value for this argument is \code{sqrt(.Machine$double.eps)}, where \code{.Machine$double.eps} is the smallest number R recognizes.}

\item{...}{other arguments (such as \code{breaks}, \code{freq}, and \code{xlim}) which are passed to the \code{hist()} function within this function. Please note that this function allows for the use of any of the primary arguments of the \code{hist()} function, but it does not support additional arguments of the \code{hist()} function accessible through its \code{...} argument (for more information, run \code{help(hist)}).}
}
\value{
This function returns a list with two elements. The first element, \code{Vertical_Axis_Limits}, contains the vertical-axis limits for the histogram; these vertical axis limits optimize the space allocated to the text box - they minimize unused plotting area and they prevent text from falling on top of the histogram bars or outside the plotting area. The second element, \code{Text_Vertical_Coordinate}, is the vertical-axis coordinate for the text box (it is assumed that the horizontal-axis coordinate for the text box will be the horizontal center of the histogram, which can be determined using \code{mean(par("usr")[1:2])}).
}
\description{
This function may be used to determine optimal vertical-axis limits of histograms when a text box is to be added above all the histogram bars and within the plotting area. Using the \code{Text} and \code{Text_Size} arguments, this function minimizes unused plotting space in the plotting window.
}
\details{
\code{Adding_a_Text_Box_at_the_Top_of_a_Histogram} may be used to determine optimal vertical-axis limits of histograms when a text box is to be added above all the histogram bars and within the plotting area. Using the \code{Text} and \code{Text_Size} arguments, this function minimizes unused plotting space in the plotting window.

This function works by using the data provided to make a histogram, extracting the vertical-axis limits of the histogram, determining how large the desired text is based on the existing plotting region, and then calculating new vertical-axis limits based on the range of the response variable and the text height. Then, these new vertical-axis limits are used to plot the data again, the text size is again determined based on this new plotting area, and another new set of vertical-axis limits are determined using the same criterion from the previous iteration. This iterative process repeats until it converges to a particular set of vertical-axis limits.
}
\examples{
set.seed(42)
Areas <- rlnorm(100)
Text <- bquote(atop("Mean:" ~ .(round(mean(Areas), 3)) ~ cm ^ 2, "Standard Deviation:" ~ .(round(sd(Areas), 3)) ~ cm ^ 2))
Text_Size <- 1.125
Plot_Margins <- c(5, 5, 4, 1)
(Output <- Adding_a_Text_Box_at_the_Top_of_a_Histogram(Areas, Text, Text_Size = Text_Size, Plot_Margins = Plot_Margins))
par(mar = Plot_Margins)
hist(Areas, ylim = Output$Vertical_Axis_Limits, xlab = expression(paste("Area (cm" ^ "2" * ")")), main = "Example Histogram With a Text Box")
text(mean(par("usr")[1:2]), Output$Text_Vertical_Coordinate, Text, cex = Text_Size)

}
\author{
David B. Moore (\email{davidblakneymoore@gmail.com})
}
