% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adding_a_Text_Box_to_a_Histogram.R
\name{Adding_a_Text_Box_to_a_Histogram}
\alias{Adding_a_Text_Box_to_a_Histogram}
\title{Adding a Text Box to a Histogram}
\usage{
Adding_a_Text_Box_to_a_Histogram(
  Values,
  Text,
  Text_Size = 1,
  Frequency_or_Density = "Frequency"
)
}
\arguments{
\item{Values}{The values to use to generate the histogram.}

\item{Text}{The text to be added to the histogram. This argument may be of class \code{'character'}, \code{'language'}, \code{'expression'}, or \code{'call'}.}

\item{Text_Size}{The size of the desired text in the text box. The default for this argument is \code{1}.}

\item{Frequency_or_Density}{The vertical-axis variable of the histogram. The default to this argument is \code{"Frequency"}.}
}
\value{
This function returns a list with two elements. The first element, \code{Vertical_Axis_Limits}, contains the vertical-axis limits for the histogram; these vertical axis limits optimize the space allocated to the text box - they minimize unused plotting area and they prevent text from falling on top of the histogram bars or outside the plotting area. The second element, \code{Text_Vertical_Coordinate}, is the vertical-axis coordinate for the text box (it is assumed that the horizontal-axis coordinate for the text box will be the horizontal center of the histogram, which can be determined using \code{mean(par("usr")[1:2])}).
}
\description{
This function may be used to determine optimal vertical-axis limits of histograms when a text box is to be added above all the histogram bars and within the plotting area. Using the \code{Text} and \code{Text_Size} arguments, this function minimizes unused plotting space in the plotting window.
}
\details{
\code{Adding_a_Text_Box_to_a_Histogram} may be used to determine optimal vertical-axis limits of histograms when a text box is to be added above all the histogram bars and within the plotting area. Using the \code{Text} and \code{Text_Size} arguments, this function minimizes unused plotting space in the plotting window.
}
\examples{
set.seed(42)
Values <- rlnorm(100)
Text <- bquote(atop("Mean:" ~ .(round(mean(Values), 3)) ~ cm ^ 2, "Standard Deviation:" ~ .(round(sd(Values), 3)) ~ cm ^ 2))
Text_Size <- 1.25
Frequency_or_Density <- "Density" # This argument may be either "Frequency" or "Density" ("Frequency" is the default)
(Output <- Adding_a_Text_Box_to_a_Histogram(Values, Text, Text_Size = 1.125, Frequency_or_Density = "Density"))
hist(Values, freq = F, ylim = Output$Vertical_Axis_Limits)
text(mean(par("usr")[1:2]), Output$Text_Vertical_Coordinate, Text)

}
\author{
David B. Moore (\email{davidblakneymoore@gmail.com})
}
