% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Adding_a_Text_Box_to_the_Top_or_Bottom_of_a_Plot.R
\name{Adding_a_Text_Box_to_the_Top_or_Bottom_of_a_Plot}
\alias{Adding_a_Text_Box_to_the_Top_or_Bottom_of_a_Plot}
\title{Adding a Text Box to the Top or Bottom of a Plot}
\usage{
Adding_a_Text_Box_to_the_Top_or_Bottom_of_a_Plot(
  Horizontal_Coordinates,
  Vertical_Coordinates,
  Text,
  Text_Box_Location = "Top",
  Text_Size = 1,
  Plot_Margins = c(5, 4, 4, 2) + 0.1,
  Tolerance = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{Horizontal_Coordinates}{The horizontal-axis values of the plot.}

\item{Vertical_Coordinates}{The vertical-axis values of the plot.}

\item{Text}{The text to be added to the histogram. This argument may be of class \code{'character'}, \code{'language'}, \code{'expression'}, or \code{'call'}.}

\item{Text_Box_Location}{The position of the text box. The possible values for this argument are \code{"Top"} and \code{"Bottom"}. The default for this argument is \code{"Top"}.}

\item{Text_Size}{The size of the desired text in the text box. The default for this argument is \code{1}.}

\item{Plot_Margins}{The plot margins, which are passed to the \code{mar} argument of the \code{par()} function. The default value for this argument is the default plot margins, \code{c(5, 4, 4, 2) + 0.1}.}

\item{Tolerance}{How close together vertical-axis limits from consecutive iterations must be for the algorithm to converge to a solution. The default value for this argument is \code{sqrt(.Machine$double.eps)}, where \code{.Machine$double.eps} is the smallest number R recognizes.}
}
\value{
This function returns a list with three elements. The first element, \code{Vertical_Axis_Limits}, contains the vertical-axis limits for the plot; these vertical-axis limits optimize the space allocated to the text box - they minimize unused plotting area and they prevent text from falling on top of the points or outside of the plotting area. The second element, \code{Text_Vertical_Coordinate}, is the vertical-axis coordinate for the text box (it is assumed that the horizontal-axis coordinate for the text box will be the horizontal center of the histogram, which can be determined using \code{mean(par("usr")[1:2])}). The third element, \code{Text_Box_Border_Coordinates}, contains both the horizontal-axis and vertical-axis coordinates of the border of the text box, which may be plotted using the \code{polygon()} function if it's desired.
}
\description{
This function may be used to determine optimal vertical-axis limits of plots when a text box is to be added at the top or the bottom of the plotting region. This function minimizes unused plotting space in the plotting window and maximizes the amount of space allocated toward the text and the points or lines.
}
\details{
\code{Adding_a_Text_Box_to_the_Top_or_Bottom_of_a_Plot} may be used to determine optimal vertical-axis limits of plots when a text box is to be added at the top or the bottom of the plotting region. This function minimizes unused plotting space in the plotting window and maximizes the amount of space allocated toward the text and the points or lines.

This function works by plotting the provided data, extracting the vertical-axis limits of the plot, determining how large the desired text is based on the existing plotting region, and then calculating new vertical-axis limits based on the range of the response variable and the text height. Then, these new vertical-axis limits are used to make another plot using the data, the text size is again determined based on this new plotting area, and another new set of vertical-axis limits are determined using the same criterion from the previous iteration. This iterative process repeats until it converges to a particular set of vertical-axis limits.
}
\examples{
set.seed(41)
Horizontal_Coordinates <- rnorm(10)
Vertical_Coordinates <- rnorm(10)
Text <- bquote(atop("Predictor Variable Mean:" ~ .(round(mean(Horizontal_Coordinates), 3)) ~ cm ^ 2, "Response Variable Mean:" ~ .(round(mean(Vertical_Coordinates), 3)) ~ cm ^ 2))
Text_Size <- 1.125
Plot_Margins <- c(5, 5, 4, 1)
Add_a_Text_Box_Border <- T
(Output <- Adding_a_Text_Box_to_the_Top_or_Bottom_of_a_Plot(Horizontal_Coordinates, Vertical_Coordinates, Text, Text_Size = Text_Size, Plot_Margins = Plot_Margins))
par(mar = Plot_Margins)
plot(Horizontal_Coordinates, Vertical_Coordinates, ylim = Output$Vertical_Axis_Limits, xlab = expression(paste("Area 1 (cm" ^ "2" * ")")), , ylab = expression(paste("Area 2 (cm" ^ "2" * ")")), main = "Example Plot With a Text Box")
text(mean(par("usr")[1:2]), Output$Text_Vertical_Coordinate, Text, cex = Text_Size)
if (Add_a_Text_Box_Border == T) {
  polygon(Output$Text_Box_Border_Coordinates$Horizontal_Axis_Coordinates, Output$Text_Box_Border_Coordinates$Vertical_Axis_Coordinates)
}

}
\author{
David B. Moore (\email{davidblakneymoore@gmail.com})
}
