% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculating_Standardized_Moments.R
\name{Calculating_Standardized_Moments}
\alias{Calculating_Standardized_Moments}
\title{Calculating_Standardized_Moments}
\usage{
Calculating_Standardized_Moments(Values, Moments)
}
\arguments{
\item{Values}{an R object containing the observations for which the standardized moments will be calculated.}

\item{Moments}{the moments (in integer form) desired.}
}
\value{
This function returns a vector of the standardized moments desired.
}
\description{
This function calculates, for a set of observations, \href{https://en.wikipedia.org/wiki/Standardized_moment}{standardized moments}. It assumes that the observations are a population, not a sample, and therefore it does not perform any bias corrections. This function relies on the \code{Ordinalizing_Integers} function which is also found in the \code{DBM.functions} package.
}
\details{
\code{Calculating_Standardized_Moments} calculates, for a set of observations, \href{https://en.wikipedia.org/wiki/Standardized_moment}{standardized moments}. It assumes that the observations are a population, not a sample, and therefore it does not perform any bias corrections.
}
\examples{
Calculating_Standardized_Moments(rlnorm(10), c(1:2, 4:7))

}
\author{
David B. Moore (\email{davidblakneymoore@gmail.com})
}
