% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checking_Layout_Matrix_Viability.R
\name{Checking_Layout_Matrix_Viability}
\alias{Checking_Layout_Matrix_Viability}
\title{Checking Layout Matrix Viability}
\usage{
Checking_Layout_Matrix_Viability(
  Matrix,
  Should_Skipping_Plot_Numbers_Be_Allowed = FALSE
)
}
\arguments{
\item{Matrix}{the matrix whose viability as a layout matrix you wish to check.}

\item{Should_Skipping_Plot_Numbers_Be_Allowed}{whether layout matrices that skip plot numbers should be allowed. If skipping numbers is allowed, some plots will not appear in the final figure, but it could be intentional - perhaps a throwaway plot was made for some other purpose. The default to this argument is \code{FALSE}.}
}
\value{
This function returns a list. The first list element, entitled \code{"Numbering_Stuff"}, contains information pertaining to if the matrix ???
}
\description{
This function determines if a potential layout matrix (for plotting) is viable by determining if any plot numbers are skipped and if all potential plotting regions are single, contiguous, perfect rectangles.
}
\details{
\code{Checking_Layout_Matrix_Viability} determines if a potential layout matrix (for plotting) is viable by determining if all potential plotting regions are rectangular and contiguous.
}
\examples{
(Example_Matrix_1 <- matrix(c(1, 1, 1, 1, 2, 2, 3, 3, 4, 5, 6, 7), byrow = T, ncol = 4))
(Example_Matrix_2 <- matrix(c(1, 1, 1, 2, 0, 2, 3, 4, 5), byrow = T, ncol = 3))
(Example_Matrix_3 <- matrix(c(1, 1, 1, 1, 1, 2, 3, 3, 3, 4, 5, 6), byrow = T, ncol = 3))
(Example_Matrix_4 <- matrix(c(1, 1, 1, 1, 1, 1, 3, 3, 3, 4, 5, 6), byrow = T, ncol = 3))
Checking_Layout_Matrix_Viability(Example_Matrix_1)
Checking_Layout_Matrix_Viability(Example_Matrix_2)
Checking_Layout_Matrix_Viability(Example_Matrix_3)
Checking_Layout_Matrix_Viability(Example_Matrix_4)
Checking_Layout_Matrix_Viability(Example_Matrix_4, Should_Skipping_Plot_Numbers_Be_Allowed = TRUE)

}
\author{
David B. Moore (\email{davidblakneymoore@gmail.com})
}
