% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Comparing_Multiple_Independent_Correlation_Coefficients.R
\name{Comparing_Multiple_Independent_Correlation_Coefficients}
\alias{Comparing_Multiple_Independent_Correlation_Coefficients}
\title{Comparing Multiple Independent Correlation Coefficients}
\usage{
Comparing_Multiple_Independent_Correlation_Coefficients(
  Correlation_Coefficients,
  Numbers_of_Observations,
  Identifiers,
  Data_Frame,
  Alpha = 0.05,
  Control_for_Experimentwise_Error = TRUE
)
}
\arguments{
\item{Correlation_Coefficients}{a numeric vector containing the correlation coefficients to be analyzed.}

\item{Numbers_of_Observations}{a numeric or an integer vector containing the numbers of observations that went in to each of the corresponding correlation coefficients.}

\item{Identifiers}{a character or a factor vector containing names to identify each corresponding correlation coefficient.}

\item{Data_Frame}{an optional data frame to include such that column names can be supplied for the first three arguments (the data frame that these columns are from should be provided for the \code{Data_Frame} argument).}

\item{Alpha}{a value of alpha against which significance can be tested (the default is \code{0.05}).}

\item{Control_for_Experimentwise_Error}{a logical argument specifying whether or not this function should give conservative estimates (by holding the experimentwise error rate at the given value of \code{Alpha}) or liberal estimates (by using the given value of \code{Alpha} for each pairwise comparison). The default, \code{TRUE}, holds the experimentwise error rate at \code{Alpha} and calculates the comparisonwise error rate based on the number of pairwise comparisons. Since this procedure assumes independence of correlations, the Šidák correction, which is used when there is independence, is used (Šidák, 1967).}
}
\value{
This function returns a data frame containing p values for each pairwise correlation coefficient comparison, a data frame containing separation lettering, and a data frame containing metadata.
}
\description{
This function determines if correlation coefficients from independent correlations are statistically significantly different (Levy, 1977).
}
\details{
\code{Comparing_Multiple_Independent_Correlation_Coefficients} compares multiple correlation coefficients from independent correlations (Levy, 1977). When independence of correlations cannot be assumed, please consider using a different procedure. This function only uses \code{base} R functions, but it was heavily inspired by the \code{agricolae} package, particularly the \code{orderPvalue()} and \code{lastC()} functions (de Mendiburu, 2017).
}
\examples{
# Generate Some Made-up Data
Made_Up_Data_for_the_Comparing_Multiple_Independent_Correlation_Coefficients_Function <- structure(list(Name = c("Correlation A", "Correlation B", "Correlation C", "Correlation D", "Correlation E"), Coefficient_of_Correlation = c(-0.339, 0.307, 0.919, -0.679, -0.495), Sample_Size = c(42L, 10L, 46L, 98L, 63L)), class = "data.frame", row.names = c(NA, -5L))

# Compare the Correlation Coefficients
(DBM.functions::Comparing_Multiple_Independent_Correlation_Coefficients(Coefficient_of_Correlation, Sample_Size, Name, Made_Up_Data_for_the_Comparing_Multiple_Independent_Correlation_Coefficients_Function))

}
\references{
de Mendiburu, F. 2017. agricolae: Statistical Procedures for Agricultural Research. R package version 1.2-8. \url{https://CRAN.R-project.org/package=agricolae}.

Levy, K.J. 1977. Pairwise comparisons involving unequal sample sizes associated with correlations, proportions or variances. Br. J. Math. Stat. Psychol. 30:137-139.

Šidák, Z.K. 1967. Rectangular Confidence Regions for the Means of Multivariate Normal Distributions. J. Am. Stat. Assoc. 62: 626–633.
}
\author{
David B. Moore (\email{davidblakneymoore@gmail.com})
}
