% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Finding_the_Full_Diagonal_of_a_Matrix.R
\name{Finding_the_Full_Diagonal_of_a_Matrix}
\alias{Finding_the_Full_Diagonal_of_a_Matrix}
\title{Finding the Full Diagonal of a Matrix}
\usage{
Finding_the_Full_Diagonal_of_a_Matrix(Matrix)
}
\arguments{
\item{Matrix}{the matrix for which the full diagonal will be calculated.}
}
\value{
This function returns the vector of matrix entries that fall on the full diagonal of a matrix (\code{Full_Diagonal}), a matrix containing the distances of each entry to the full diagonal (\code{Distances_From_the_Full_Diagonal}), and the original matrix (\code{Original_Matrix}).
}
\description{
This function finds the full diagonal of a matrix. A full diagonal of a matrix is the diagonal of the matrix that is the line segment that starts at the entry in the first row and the first column of the matrix and that ends at the entry in the last row and the last column of the matrix.
}
\details{
\code{Finding_the_Full_Diagonal_of_a_Matrix} finds the full diagonal of a matrix. A full diagonal of a matrix is the diagonal of the matrix that is the line segment that starts at the entry in the first row and the first column of the matrix and that ends at the entry in the last row and the last column of the matrix. If the numbers of rows and columns are coprime, the only two matrix entries that fall perfectly along the full diagonal are its two endpoints; when the numbers of rows and coliumns are not coprime, there will be more than two elements of the full diagonal. For square matrices, the \href{https://en.wikipedia.org/wiki/Main_diagonal}{main diagonal} and the full diagonal are identical, but for non-square matrices, the main diagonal and the full diagonal are not identical.
}
\examples{
# An example where the number of columns is greater than the number of rows
Matrix_1 <- matrix(1:24, nrow = 4)
Finding_the_Full_Diagonal_of_a_Matrix(Matrix_1)

# An example where the number of rows is greater than the number of columns
Matrix_2 <- matrix(1:24, ncol = 4)
Finding_the_Full_Diagonal_of_a_Matrix(Matrix_2)

# A non-numeric example where the number of columns equals the number of rows
Matrix_3 <- matrix(LETTERS[1:16], ncol = 4)
Finding_the_Full_Diagonal_of_a_Matrix(Matrix_3)

# An example where the matrix only has one row
Matrix_4 <- matrix(letters, nrow = 1)
Finding_the_Full_Diagonal_of_a_Matrix(Matrix_4)

# An example where the matrix only has one column
Matrix_5 <- matrix(letters, ncol = 1)
Finding_the_Full_Diagonal_of_a_Matrix(Matrix_5)

# An example where the matrix has no rows and no columns
Matrix_6 <- matrix(NA, nrow = 0, ncol = 0)
Finding_the_Full_Diagonal_of_a_Matrix(Matrix_6)

}
\author{
David B. Moore (\email{davidblakneymoore@gmail.com})
}
