% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Generating_Figures_Containing_a_Bipartite_Interaction_Matrix_Visualization_Plot_and_Corresponding_Upper_and_Lower_Trophic_Level_Phenology_Plots.R
\name{Generating_Figures_Containing_a_Bipartite_Interaction_Matrix_Visualization_Plot_and_Corresponding_Higher_and_Lower_Trophic_Level_Phenology_Plots}
\alias{Generating_Figures_Containing_a_Bipartite_Interaction_Matrix_Visualization_Plot_and_Corresponding_Higher_and_Lower_Trophic_Level_Phenology_Plots}
\title{Generating Figures Containing a Bipartite Interaction Matrix Visualization Plot and Corresponding Higher and Lower Trophic Level Phenology Plots}
\usage{
Generating_Figures_Containing_a_Bipartite_Interaction_Matrix_Visualization_Plot_and_Corresponding_Higher_and_Lower_Trophic_Level_Phenology_Plots(
  Data_Frame,
  Higher_Trophic_Level_Species,
  Lower_Trophic_Level_Species,
  Date,
  Overall_Plot_Title = NULL,
  Lower_Trophic_Level_Plot_Title = "Lower Trophic Level",
  Higher_Trophic_Level_Plot_Title = "Higher Trophic Level",
  Prioritize_Interaction_Strength_Over_Alphabetization_in_Species_Layout = FALSE,
  Lower_Trophic_Level_Color = "darkgreen",
  Higher_Trophic_Level_Color = "darkred",
  Interaction_Color = "black",
  Method_for_Aggregating = "Sum",
  Include_Scale_Bars = TRUE,
  Scale_Bar_Factor = (1/3),
  Approximate_Proportion_of_Lower_Trophic_Level_Species_to_Pool = 0,
  Approximate_Proportion_of_Higher_Trophic_Level_Species_to_Pool = 0,
  Lower_Trophic_Level_Species_Names_Data_Frame = NULL,
  Higher_Trophic_Level_Species_Names_Data_Frame = NULL,
  Relative_Plot_Widths = c(11, 4, 10, 4, 11),
  Relative_Plot_Heights = c(rep(1, ifelse(is.null(Overall_Plot_Title), 1, 2)),
    ifelse(is.null(Overall_Plot_Title), 9, 8)),
  Already_Pooled_Lower_Trophic_Level_Species_Name = NULL,
  Already_Pooled_Higher_Trophic_Level_Species_Name = NULL,
  Gap_Size_Constant = 0.2,
  Relative_Bar_Width = 0.2,
  Horizontal_Scale_Text_Shift_Constant = 0.1,
  Vertical_Scale_Title_Shift_Constant = 0.1,
  Scale_Bar_Title_Font_Size = 1,
  Scale_Bar_Text_Font_Size = 1,
  Overall_Plot_Title_Font_Size = 3.25,
  Trophic_Level_Plot_Title_Font_Size = 2.5,
  Phenology_and_Bipartite_Plot_Title_Font_Size = 1.5,
  Species_Name_Font_Size = 1.5,
  Horizontal_Axis_Title_Font_Size = 1,
  Horizontal_Axis_Dates_Font_Size = 1,
  Left_and_Right_Figure_Margins = 9,
  Between_Title_Plots_and_Other_Plots_Margin = 1,
  Bottom_Margin = 8
)
}
\arguments{
\item{Data_Frame}{a data frame containing information about the number of interactions between the higher- and lower-trophic-level species, the higher-trophic-level species, the lower-trophic-level species, the dates, and the factors that will be used to split the data to generate multiple panels for the resulting figure. This data frame must contain a column of dates (in \code{"Date"} format), a column of lower-trophic-level species (in \code{"character"} format), and one or more columns of higher-trophic-level species (containing the names of these species as the column names and \code{"numeric"} entries that represent the numbers of interactions between each particular higher-trophic-level species and the corresponding lower-trophic-level species). This argument must be provided for the function to work.}

\item{Higher_Trophic_Level_Species}{a \code{"character"} vector containing the names of each of the higher-trophic-level species in the experiment. These names must also be column names in the \code{Data_Frame} data frame; the entries of these columns will be \code{"numeric"} and will represent the numbers of interactions between higher- and lower-trophic-level species. This argument must be provided for the function to work.}

\item{Lower_Trophic_Level_Species}{a \code{"character"} vector of length \code{1} containing the name of the column in the \code{Data_Frame} data frame where the lower-trophic-level-species names are stored. This argument must be provided for the function to work.}

\item{Date}{a \code{"character"} vector of length \code{1} containing the name of the column in the \code{Data_Frame} data frame where the dates are stored. This argument must be provided for the function to work.}

\item{Overall_Plot_Title}{the overall figure title. The default value, \code{NULL}, means that an overall figure title will be omitted.}

\item{Lower_Trophic_Level_Plot_Title}{a name for the lower trophic level, such as \code{"Plants"} or \code{"Prey"}. The default value, \code{"Lower Trophic Level"}, is a generic title that can be used in any figure.}

\item{Higher_Trophic_Level_Plot_Title}{a name for the higher trophic level, such as \code{"Pollinators"} or \code{"Predators"}. The default value, \code{"Higher Trophic Level"}, is a generic title that can be used in any figure.}

\item{Prioritize_Interaction_Strength_Over_Alphabetization_in_Species_Layout}{whether species should be arranged such that species from different trophic levels that interact more frequently are closer together and species that interact less frequently are farther apart on the bipartite interaction visualization plot. If putting frequently interacting species closer together and infrequently interacting species farther apart is prioritized, the species will probably not be arranged alphabetically in the resulting figure. The default value, \code{FALSE}, causes the species from both trophic levels to be arranged alphabetically.}

\item{Lower_Trophic_Level_Color}{the color of the violin-like phenology plots and the bars (from the bipartite interaction visualization plot) for the lower-trophic-level species. The default value to this argument is \code{"darkgreen"}.}

\item{Higher_Trophic_Level_Color}{the color of the violin-like phenology plots and the bars (from the bipartite interaction visualization plot) for the higher-trophic-level species. The default value to this argument is \code{"darkred"}.}

\item{Interaction_Color}{the color of the edges (from the bipartite interaction visualization plot) for the interactions between the lower-trophic-level species and the higher-trophic-level species. The default value to this argument is \code{"black"}.}

\item{Include_Scale_Bars}{whether vertical scale bars should be included on the left and right sides of the figure. The default value, \code{TRUE}, causes scale bars to be included.}

\item{Scale_Bar_Factor}{a number between 0 and 1 (exclusive) that specifies approximately how tall the scale bar will be relative to the vertical axis range. The default value for this argument, \code{0.25}, makes it so that the scale bar will take up as close to 25 \% of the vertical axis' range as possible.}

\item{Approximate_Proportion_of_Lower_Trophic_Level_Species_to_Pool}{the approximate proportion of lower-trophic-level species to pool into a 'other species' group. First, lower-trophic-level species are ordered by their total numbers of occurrences, and then this argument is used to determine the threshold for inclusion in the figure. Species that do not meet this threshold (that have low numbers of occurrences) are pooled into an 'other species' group. The default value, \code{0}, means that no species will be pooled.}

\item{Approximate_Proportion_of_Higher_Trophic_Level_Species_to_Pool}{the approximate proportion of higher-trophic-level species to pool into a 'other species' group. First, higher-trophic-level species are ordered by their total numbers of occurrences, and then this argument is used to determine the threshold for inclusion in the figure. Species that do not meet this threshold (that have low numbers of occurrences) are pooled into an 'other species' group. The default value, \code{0}, means that no species will be pooled.}

\item{Lower_Trophic_Level_Species_Names_Data_Frame}{a data frame (of class \code{"data.frame"}) that contains two columns - one column, which must be named \code{"Row_Name"}, must have entries that identically match the names of the rows of the matrices (which represent the lower-trophic-level species), and the other column, which must be named \code{"Names_for_Plotting"}, will have entries that will appear as the (potentially stylized, using the \code{expression()} function, for example) lower-trophic-level species names in the resulting figure.}

\item{Higher_Trophic_Level_Species_Names_Data_Frame}{a data frame (of class \code{"data.frame"}) that contains two columns - one column, which must be named \code{"Column_Name"}, must have entries that identically match the names of the columns of the matrices (which represent the higher-trophic-level species), and the other column, which must be named \code{"Names_for_Plotting"}, will have entries that will appear as the (potentially stylized, using the \code{expression()} function, for example) higher-trophic-level species names in the resulting figure.}

\item{Relative_Plot_Widths}{the relative widths of the lower-trophic-level violin-like phenology plots, the plot containing lower-trophic-level species names, the bipartite interaction visualization plot, the plot containing the higher-trophic-level species names, and the higher-trophic-level violin-like phenology plots. This argument must be a numeric vector containing 5 positive numbers, and it must be the same forward and backward to ensure that the resulting figure is symmetric. The default value to this argument is \code{c(11, 4, 10, 4, 11)}.}

\item{Relative_Plot_Heights}{the relative heights of the plots containing the titles and all the other plots. This argument must be a numeric vector and contain two (if an overall figure title is not to be included) or three (if an overall figure title is to be included) positive numbers. The default value to this argument, \code{c(rep(1, ifelse(is.null(Overall_Plot_Title), 1, 2)), ifelse(is.null(Overall_Plot_Title), 9, 8))}, means that when there is no overall figure title, the heights will be \code{c(1, 9)}, and when there is an overall figure title, the heights will be \code{c(1, 1, 8)}.}

\item{Already_Pooled_Lower_Trophic_Level_Species_Name}{the row name representing already-pooled species (such as miscellaneous species or other species), if there is a row of this nature present. The default value, \code{NULL}, assumes that there are no already-pooled rows. This pooled group will be placed at the bottom of the figure, or it will be pooled with other rows if an appropriate 'Approximate_Proportion_of_Lower_Trophic_Level_Species_to_Pool' argument is provided.}

\item{Already_Pooled_Higher_Trophic_Level_Species_Name}{the column name representing already-pooled species (such as miscellaneous species or other species), if there is a column of this nature present. The default value, \code{NULL}, assumes that there are no already-pooled columns. This pooled group will be placed at the bottom of the figure, or it will be pooled with other columns if an appropriate 'Approximate_Proportion_of_Higher_Trophic_Level_Species_to_Pool' argument is provided.}

\item{Gap_Size_Constant}{a nonnegative number that specifies how much space is allocated between bars vertically (within trophic levels) in the bipartite interaction visualization plot, where each bar represents a species. Using higher numbers for this argument will lead to more space between species' bars. The default value, \code{0.2}, specifies that 20 \% of the cumulative bar heights (without any space) will be divided up equally between the bars in this plot.}

\item{Relative_Bar_Width}{a number between \code{0} and \code{1} (inclusive) that specifies how wide the bars representing species are in the bipartite interaction visualization plot. The default value, \code{0.2}, specifies that 20 \% of the plotting areas' widths are to be taken up by the bars for both trophic levels combined. The remaining 80 \% of the space will be taken up by the interactions (the edges).}

\item{Horizontal_Scale_Text_Shift_Constant}{a positive number specifying the relative horizontal distance (relative to the overall horizontal axis range) the scale bar text will be from the scale bar. The default value to this argument is \code{0.1}.}

\item{Vertical_Scale_Title_Shift_Constant}{a positive number specifying the relative vertical distance (relative to the overall vertical axis range) the scale bar title will be from the top of the scale bar. The default value to this argument is \code{0.1}.}

\item{Scale_Bar_Title_Font_Size}{the relative size of the scale bar titles. This argument must be a single positive number. The default value to this argument is \code{1}.}

\item{Scale_Bar_Text_Font_Size}{the relative size of the scale bar text. This argument must be a single positive number. The default value to this argument is \code{1}.}

\item{Overall_Plot_Title_Font_Size}{the relative size of the overall figure title. This argument must be a single positive number. The default value to this argument is \code{3.25}.}

\item{Trophic_Level_Plot_Title_Font_Size}{the relative size of the titles for the lower-trophic-level and higher-trophic-level title plots. This argument must be a single positive number. The default value to this argument is \code{2.5}.}

\item{Phenology_and_Bipartite_Plot_Title_Font_Size}{the relative size of the titles of the lower-trophic-level violin-like phenology plots, the plot containing lower-trophic-level species names, the bipartite interaction visualization plot, the plot containing the higher-trophic-level species names, and the higher-trophic-level violin-like phenology plots. This argument must be a single positive number. The default value to this argument is \code{1.5}.}

\item{Species_Name_Font_Size}{the relative size of the species names for the lower-trophic-level violin-like phenology plots and the higher-trophic-level violin-like phenology plots. This argument must be a single positive number. The default value to this argument is \code{1.5}.}

\item{Horizontal_Axis_Title_Font_Size}{the relative size of the horizontal axis titles for the lower-trophic-level violin-like phenology plots and the higher-trophic-level violin-like phenology plots. This argument must be a single positive number. The default value to this argument is \code{1}.}

\item{Horizontal_Axis_Dates_Font_Size}{the relative size of the dates on the horizontal axes of the lower-trophic-level violin-like phenology plots and higher-trophic-level violin-like phenology plots. This argument must be a single positive number. The default value to this argument is \code{1}.}

\item{Left_and_Right_Figure_Margins}{the relative distance between the left figure margin and the left edge of the lower-trophic-level violin-like phenology plots, and the relative distance between the right figure margin and the right edge of the higher-trophic-level violin-like phenology plots. This argument must be a single positive number. The default value to this argument is \code{9}.}

\item{Between_Title_Plots_and_Other_Plots_Margin}{the relative distance between the bottom of the title plots (which are on the top row of the figure) and the top of the plots on the bottom row of the figure. This argument must be a single positive number. The default value to this argument is \code{1}.}

\item{Bottom_Margin}{the relative distance between the edges of the plots on the bottom row of the figure and the lower figure margin. This argument must be a single positive number. The default value to this argument is \code{8}.}
}
\value{
This function returns a figure with violin-like phenology plots showing the abundances of species from the lower trophic levels over time on the left side of the figure and violin-like phenology plots showing the abundances of species from the higher trophic levels over time on the right side of the figure. In the middle of these two is a bipartite interaction matrix visualization which shows how the two trophic levels interacted.
}
\description{
This function generates violin-like phenology plots that depict the relative abundances of lower-trophic-level and higher-trophic-level species over time. It also generates bipartite interaction matrix visualization plots which are positioned centrally between the two sets of phenology plots.
}
\details{
Unfortunately, this function can't handle missing values at this point!

\code{Generating_Figures_Containing_a_Bipartite_Interaction_Matrix_Visualization_Plot_and_Corresponding_Higher_and_Lower_Trophic_Level_Phenology_Plots} generates violin-like phenology plots that depict the relative abundances of lower-trophic-level and higher-trophic-level species over time. It also generates bipartite interaction matrix visualization plots which are positioned centrally between the two sets of phenology plots.

In bipartite parlance, 'nodes' refer to species (from either trophic level) and 'edges' refer to interactions between these species. Bipartite networks are special because species interact only across, and not within, trophic levels.

This function was heavily inspired by a manuscript that contained a figure very similar to the one this function returns (Russo et al., 2013). It was also heavily inspired by my colleague Isaac Ativor's work in plant-pollinator interactions.

This function relies on the \link[DBM.functions]{Attempting_to_Diagonalize_a_Matrix_Based_on_Its_Full_Diagonal} function and the \link[DBM.functions]{Making_Reasonable_Scale_Bars} function, which are both also found in the \code{DBM.functions} package.
}
\examples{
# Generate Made-Up Data
Data_Frame_1 <- structure(list(Date = structure(c(18357, 18357, 18357, 18357, 18392, 18392, 18392, 18392, 18422, 18422, 18422, 18422, 18456, 18456, 18456, 18456), class = "Date"), Lower_Trophic_Level_Species = c("Species a", "Species b", "Species c", "Species d", "Species a", "Species b", "Species c", "Species d", "Species a", "Species b", "Species c", "Species d", "Species a", "Species b", "Species c", "Species d"), `Species A` = c(8, 7, 12, 2, 8, 1, 3, 0, 2, 1, 4, 0, 1, 0, 3, 1), `Species B` = c(9, 11, 6, 24, 3, 1, 4, 7, 2, 1, 16, 3, 1, 2, 0, 2), `Species C` = c(6, 5, 2, 8, 31, 10, 1, 1, 3, 1, 5, 0, 4, 1, 1, 0), `Species D` = c(9, 6, 8, 9, 2, 10, 0, 4, 2, 1, 0, 0, 1, 3, 0, 0), `Species E` = c(6, 1, 4, 2, 1, 1, 0, 3, 2, 2, 1, 0, 0, 0, 0, 1), `Species F` = c(12, 5, 13, 5, 0, 0, 3, 8, 1, 1, 2, 4, 1, 2, 1, 1), `Species G` = c(10, 11, 5, 1, 1, 0, 1, 1, 1, 4, 1, 0, 0, 2, 0, 0)), row.names = c(NA, -16L), class = "data.frame")
Higher_Trophic_Level_Species_1 <- c("Species A", "Species B", "Species C", "Species D", "Species E", "Species F", "Species G")
Lower_Trophic_Level_Species_1 <- "Lower_Trophic_Level_Species"
Date_1 <- "Date"
Lower_Trophic_Level_Species_Names_Data_Frame_1 <- data.frame(Lower_Trophic_Level_Species_Name = paste("Species", letters[seq_len(4)]), Names_for_Plotting = c("expression(paste(italic('Lotus corniculatus')))", "expression(paste(italic('Medicago sativa')))", "expression(paste(italic('Pisum'), ' sp.'))", "expression(paste(italic('Trifolium'), ' spp.'))"))
Higher_Trophic_Level_Species_Names_Data_Frame_1 <- data.frame(Column_Name = paste("Species", LETTERS[seq_len(7)]), Names_for_Plotting = c("expression(paste(italic('Apis mellifera')))", "expression(paste(italic('Bombus'), ' spp.'))", "expression(paste('Coleoptera spp.'))", "expression(paste(italic('Dolichiovespula'), ' spp.'))", "expression(paste('Lepidoptera spp.'))", "expression(paste(italic('Vespa'), ' spp.'))", "expression(paste(italic('Xylocopa'), ' spp.'))"))
Data_Frame_2 <- structure(list(Date = structure(c(18357, 18357, 18357, 18357, 18392, 18392, 18392, 18392, 18422, 18422, 18422, 18422, 18456, 18456, 18456, 18456), class = "Date"), Lower_Trophic_Level_Species = c("Species a", "Species b", "Species c", "Species d", "Species a", "Species b", "Species c", "Species d", "Species a", "Species b", "Species c", "Species d", "Species a", "Species b", "Species c", "Species d"), `Species A` = c(3, 3, 9, 14, 3, 2, 1, 57, 1, 0, 0, 2, 1, 2, 1, 1), `Species B` = c(6, 6, 8, 5, 3, 1, 1, 0, 1, 4, 2, 1, 1, 0, 0, 3), `Species C` = c(8, 5, 12, 5, 1, 1, 1, 1, 1, 7, 1, 2, 0, 1, 1, 4), `Species D` = c(1, 3, 5, 2, 1, 6, 3, 2, 0, 8, 7, 4, 0, 0, 6, 1), `Species E` = c(2, 7, 7, 7, 1, 1, 0, 2, 1, 1, 3, 1, 0, 1, 0, 1), `Species F` = c(5, 3, 5, 3, 4, 0, 0, 1, 1, 5, 1, 0, 1, 0, 0, 0), `Miscellaneous Species` = c(6, 11, 6, 14, 1, 0, 1, 8, 1, 3, 0, 2, 4, 1, 1, 1)), row.names = c(NA, -16L), class = "data.frame")
Higher_Trophic_Level_Species_2 <- c("Species A", "Species B", "Species C", "Species D", "Species E", "Species F", "Miscellaneous Species")
Lower_Trophic_Level_Species_2 <- "Lower_Trophic_Level_Species"
Date_2 <- "Date"
Lower_Trophic_Level_Species_2 <- "Lower_Trophic_Level_Species"

# Example Figure 1
Generating_Figures_Containing_a_Bipartite_Interaction_Matrix_Visualization_Plot_and_Corresponding_Higher_and_Lower_Trophic_Level_Phenology_Plots(Data_Frame_1, Higher_Trophic_Level_Species_1, Lower_Trophic_Level_Species_1, Date_1, Include_Scale_Bars = FALSE)

# Example Figure 2
Generating_Figures_Containing_a_Bipartite_Interaction_Matrix_Visualization_Plot_and_Corresponding_Higher_and_Lower_Trophic_Level_Phenology_Plots(Data_Frame_1, Higher_Trophic_Level_Species_1, Lower_Trophic_Level_Species_1, Date_1, Scale_Bar_Factor = (1 / 3), Prioritize_Interaction_Strength_Over_Alphabetization_in_Species_Layout = TRUE, Approximate_Proportion_of_Higher_Trophic_Level_Species_to_Pool = 0.25)

# Example Figure 3
Generating_Figures_Containing_a_Bipartite_Interaction_Matrix_Visualization_Plot_and_Corresponding_Higher_and_Lower_Trophic_Level_Phenology_Plots(Data_Frame_2, Higher_Trophic_Level_Species_2, Lower_Trophic_Level_Species_2, Date_2, Scale_Bar_Factor = (1 / 3), Already_Pooled_Higher_Trophic_Level_Species_Name = "Miscellaneous Species", Species_Name_Font_Size = 1.25)

# Example Figure 4
Generating_Figures_Containing_a_Bipartite_Interaction_Matrix_Visualization_Plot_and_Corresponding_Higher_and_Lower_Trophic_Level_Phenology_Plots(Data_Frame_1, Higher_Trophic_Level_Species_1, Lower_Trophic_Level_Species_1, Date_1, Scale_Bar_Factor = (1 / 3), Lower_Trophic_Level_Species_Names_Data_Frame = Lower_Trophic_Level_Species_Names_Data_Frame_1, Higher_Trophic_Level_Species_Names_Data_Frame = Higher_Trophic_Level_Species_Names_Data_Frame_1, Overall_Plot_Title = "Example Figure 4", Lower_Trophic_Level_Plot_Title = "Plants", Higher_Trophic_Level_Plot_Title = "Pollinators")

}
\references{
Russo, L., N. DeBarros, S. Yang, K. Shea, and D. Mortensen. 2013. Supporting crop pollinators with floral resources: network-based phenological matching. Ecol. Evol. 3:3125-3140.
}
\author{
David B. Moore (\email{davidblakneymoore@gmail.com})
}
