% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Generating_Figures_Containing_a_Bipartite_Interaction_Matrix_Visualization_Plot_and_Corresponding_Higher_and_Lower_Trophic_Level_Phenology_Plots_by_Factors.R
\name{Generating_Figures_Containing_a_Bipartite_Interaction_Matrix_Visualization_Plot_and_Corresponding_Higher_and_Lower_Trophic_Level_Phenology_Plots_by_Factors}
\alias{Generating_Figures_Containing_a_Bipartite_Interaction_Matrix_Visualization_Plot_and_Corresponding_Higher_and_Lower_Trophic_Level_Phenology_Plots_by_Factors}
\title{Generating Figures Containing a Bipartite Interaction Matrix Visualization Plot and Corresponding Higher and Lower Trophic Level Phenology Plots by Factors}
\usage{
Generating_Figures_Containing_a_Bipartite_Interaction_Matrix_Visualization_Plot_and_Corresponding_Higher_and_Lower_Trophic_Level_Phenology_Plots_by_Factors(
  Data_Frame,
  Higher_Trophic_Level_Species,
  Lower_Trophic_Level_Species,
  Date,
  Factors_to_Split_by,
  Factor_Data_for_Plotting = NULL,
  Factor_Level_Data_for_Plotting = NULL,
  Overall_Plot_Title_Factor_Arrangement = "On Different Lines",
  Lower_Trophic_Level_Plot_Title = "Lower Trophic Level",
  Higher_Trophic_Level_Plot_Title = "Higher Trophic Level",
  Prioritize_Interaction_Strength_Over_Alphabetization_in_Species_Layout = FALSE,
  Lower_Trophic_Level_Color = "darkgreen",
  Higher_Trophic_Level_Color = "darkred",
  Interaction_Color = "black",
  Method_for_Aggregating = "Sum",
  Include_Scale_Bars = TRUE,
  Scale_Bar_Factor = 0.25,
  Approximate_Proportion_of_Lower_Trophic_Level_Species_to_Pool = 0,
  Approximate_Proportion_of_Higher_Trophic_Level_Species_to_Pool = 0,
  Lower_Trophic_Level_Species_Names_Data_Frame = NULL,
  Higher_Trophic_Level_Species_Names_Data_Frame = NULL,
  Relative_Plot_Widths = c(11, 4, 10, 4, 11),
  Relative_Plot_Heights = c(1, 1, 8),
  Already_Pooled_Lower_Trophic_Level_Species_Name = NULL,
  Already_Pooled_Higher_Trophic_Level_Species_Name = NULL,
  Gap_Size_Constant = 0.2,
  Relative_Bar_Width = 0.2,
  Horizontal_Scale_Text_Shift_Constant = 0.1,
  Vertical_Scale_Title_Shift_Constant = 0.1,
  Scale_Bar_Title_Font_Size = 1,
  Scale_Bar_Text_Font_Size = 1,
  Overall_Plot_Title_Font_Size = 3.25,
  Trophic_Level_Plot_Title_Font_Size = 2.5,
  Phenology_and_Bipartite_Plot_Title_Font_Size = 1.5,
  Species_Name_Font_Size = 1.5,
  Horizontal_Axis_Title_Font_Size = 1,
  Horizontal_Axis_Dates_Font_Size = 1,
  Left_and_Right_Figure_Margins = 9,
  Between_Title_Plots_and_Other_Plots_Margin = 1,
  Bottom_Margin = 8
)
}
\arguments{
\item{Data_Frame}{a data frame containing information about the number of interactions between the higher- and lower-trophic-level species, the higher-trophic-level species, the lower-trophic-level species, the dates, and the factors that will be used to split the data to generate multiple panels for the resulting figure. This data frame must contain a column of dates (in \code{"Date"} format), a column of lower-trophic-level species (in \code{"character"} format), one or more columns of higher-trophic-level species (containing the names of these species as the column names and \code{"numeric"} entries that represent the numbers of interactions between each particular higher-trophic-level species and the corresponding lower-trophic-level species), and columns for each of the \code{Factors_to_Split_by} (each of these columns will have \code{"character"} entries that will correspond to factors of an experiment; multifactorial experimental designs must be perfectly balanced, meaning that the levels of each factor must be perfectly crossed with all of the levels of all of the other factors). This argument must be provided for the function to work.}

\item{Higher_Trophic_Level_Species}{a \code{"character"} vector containing the names of each of the higher-trophic-level species in the experiment. These names must also be column names in the \code{Data_Frame} data frame; the entries of these columns will be \code{"numeric"} and will represent the numbers of interactions between higher- and lower-trophic-level species. This argument must be provided for the function to work.}

\item{Lower_Trophic_Level_Species}{a \code{"character"} vector of length \code{1} containing the name of the column in the \code{Data_Frame} data frame where the lower-trophic-level-species names are stored. This argument must be provided for the function to work.}

\item{Date}{a \code{"character"} vector of length \code{1} containing the name of the column in the \code{Data_Frame} data frame where the dates are stored. This argument must be provided for the function to work.}

\item{Factors_to_Split_by}{a \code{"character"} vector containing the names of the columns in the \code{Data_Frame} data frame that correspond to the factors in the factorial experimental design. These entries of these columns represent the levels of each factor. This argument must be provided for the function to work.}

\item{Factor_Data_for_Plotting}{a data frame (of class \code{"data.frame"}) that contains two columns - one column, which must be named \code{"Factor"}, must have entries that identically match the names of the \code{Factors_to_Split_by} argument; the other column, which must be named \code{"Factor_Name_for_Plotting"}, must contain the names of the factors as the user of the function would like them to appear in the figure titles in the resulting figure. This argument is optional, and if it is omitted, the default is to use the elements of the \code{Factors_to_Split_by} argument in the figure titles in the final figure.}

\item{Factor_Level_Data_for_Plotting}{a data frame (of class \code{"data.frame"}) that contains three columns - one column, which must be named \code{"Factor"}, must have entries that identically match the names of the \code{Factors_to_Split_by} argument; one column, which must be named \code{"Level"}, must contain the levels of each factor exactly as they appear in the \code{Data_Frame} data frame in each factor's column; one column, which must be named \code{"Level_Name_for_Plotting"}, must contain the names of the factor levels as the user of the function would like them to appear in the figure titles in the resulting figure. This argument is optional, and if it is omitted, the default is to use the names of the factor levels as they appear in the \code{Data_Frame} data frame in the figure titles in the final figure.}

\item{Overall_Plot_Title_Factor_Arrangement}{for multifactorial experimental designs, how figure titles should be arranged. With multiple factors, and with each panel within the overall figure corresponding to one combination of factor levels, there will be multiple factors and factor levels appearing in the title of each panel. Should each factor and its corresponding factor level appear in the same line, separated by a semicolon, or should they appear in separate lines? The two possible values for this argument are \code{"On Different Lines"} and \code{"On the Same Lines"}, with the default value being \code{"On Different Lines"}.}

\item{Lower_Trophic_Level_Plot_Title}{a name for the lower trophic level, such as \code{"Plants"} or \code{"Prey"}. The default value, \code{"Lower Trophic Level"}, is a generic title that can be used in any figure.}

\item{Higher_Trophic_Level_Plot_Title}{a name for the higher trophic level, such as \code{"Pollinators"} or \code{"Predators"}. The default value, \code{"Higher Trophic Level"}, is a generic title that can be used in any figure.}

\item{Prioritize_Interaction_Strength_Over_Alphabetization_in_Species_Layout}{whether species should be arranged such that species from different trophic levels that interact more frequently are closer together and species that interact less frequently are farther apart on the bipartite interaction visualization plot. If putting frequently interacting species closer together and infrequently interacting species farther apart is prioritized, the species will probably not be arranged alphabetically in the resulting figure. The default value, \code{FALSE}, causes the species from both trophic levels to be arranged alphabetically.}

\item{Lower_Trophic_Level_Color}{the color of the violin-like phenology plots and the bars (from the bipartite interaction visualization plot) for the lower-trophic-level species. The default value to this argument is \code{"darkgreen"}.}

\item{Higher_Trophic_Level_Color}{the color of the violin-like phenology plots and the bars (from the bipartite interaction visualization plot) for the higher-trophic-level species. The default value to this argument is \code{"darkred"}.}

\item{Interaction_Color}{the color of the edges (from the bipartite interaction visualization plot) for the interactions between the lower-trophic-level species and the higher-trophic-level species. The default value to this argument is \code{"black"}.}

\item{Include_Scale_Bars}{whether vertical scale bars should be included on the left and right sides of the figure. The default value, \code{TRUE}, causes scale bars to be included.}

\item{Scale_Bar_Factor}{a number between 0 and 1 (exclusive) that specifies approximately how tall the scale bar will be relative to the vertical axis range. The default value for this argument, \code{0.25}, makes it so that the scale bar will take up as close to 25 \% of the vertical axis' range as possible.}

\item{Approximate_Proportion_of_Lower_Trophic_Level_Species_to_Pool}{the approximate proportion of lower-trophic-level species to pool into a 'other species' group. First, lower-trophic-level species are ordered by their total numbers of occurrences, and then this argument is used to determine the threshold for inclusion in the figure. Species that do not meet this threshold (that have low numbers of occurrences) are pooled into an 'other species' group. The default value, \code{0}, means that no species will be pooled.}

\item{Approximate_Proportion_of_Higher_Trophic_Level_Species_to_Pool}{the approximate proportion of higher-trophic-level species to pool into a 'other species' group. First, higher-trophic-level species are ordered by their total numbers of occurrences, and then this argument is used to determine the threshold for inclusion in the figure. Species that do not meet this threshold (that have low numbers of occurrences) are pooled into an 'other species' group. The default value, \code{0}, means that no species will be pooled.}

\item{Lower_Trophic_Level_Species_Names_Data_Frame}{a data frame (of class \code{"data.frame"}) that contains two columns - one column, which must be named \code{"Lower_Trophic_Level_Species_Name"}, must have entries that identically match the names of the rows of the matrices (which represent the lower-trophic-level species), and the other column, which must be named \code{"Names_for_Plotting"}, will have entries that will appear as the (potentially stylized, using the \code{expression()} function, for example) lower-trophic-level species names in the resulting figure.}

\item{Higher_Trophic_Level_Species_Names_Data_Frame}{a data frame (of class \code{"data.frame"}) that contains two columns - one column, which must be named \code{"Higher_Trophic_Level_Species_Name"}, must have entries that identically match the names of the columns of the matrices (which represent the higher-trophic-level species), and the other column, which must be named \code{"Names_for_Plotting"}, will have entries that will appear as the (potentially stylized, using the \code{expression()} function, for example) higher-trophic-level species names in the resulting figure.}

\item{Relative_Plot_Widths}{the relative widths of the lower-trophic-level violin-like phenology plots, the plot containing lower-trophic-level species names, the bipartite interaction visualization plot, the plot containing the higher-trophic-level species names, and the higher-trophic-level violin-like phenology plots. This argument must be a numeric vector containing 5 positive numbers, and it must be the same forward and backward to ensure that the resulting figure is symmetric. The default value to this argument is \code{c(11, 4, 10, 4, 11)}.}

\item{Relative_Plot_Heights}{the relative heights of the plots containing the titles and all the other plots. This argument must be a numeric vector and contain two (if an overall figure title is not to be included) or three (if an overall figure title is to be included) positive numbers. The default value to this argument, \code{c(rep(1, ifelse(is.null(Overall_Plot_Title), 1, 2)), ifelse(is.null(Overall_Plot_Title), 9, 8))}, means that when there is no overall figure title, the heights will be \code{c(1, 9)}, and when there is an overall figure title, the heights will be \code{c(1, 1, 8)}.}

\item{Already_Pooled_Lower_Trophic_Level_Species_Name}{the row name representing already-pooled species (such as miscellaneous species or other species), if there is a row of this nature present. The default value, \code{NULL}, assumes that there are no already-pooled rows. This pooled group will be placed at the bottom of the figure, or it will be pooled with other rows if an appropriate 'Approximate_Proportion_of_Lower_Trophic_Level_Species_to_Pool' argument is provided.}

\item{Already_Pooled_Higher_Trophic_Level_Species_Name}{the column name representing already-pooled species (such as miscellaneous species or other species), if there is a column of this nature present. The default value, \code{NULL}, assumes that there are no already-pooled columns. This pooled group will be placed at the bottom of the figure, or it will be pooled with other columns if an appropriate 'Approximate_Proportion_of_Higher_Trophic_Level_Species_to_Pool' argument is provided.}

\item{Gap_Size_Constant}{a nonnegative number that specifies how much space is allocated between bars vertically (within trophic levels) in the bipartite interaction visualization plot, where each bar represents a species. Using higher numbers for this argument will lead to more space between species' bars. The default value, \code{0.2}, specifies that 20 \% of the cumulative bar heights (without any space) will be divided up equally between the bars in this plot.}

\item{Relative_Bar_Width}{a number between \code{0} and \code{1} (inclusive) that specifies how wide the bars representing species are in the bipartite interaction visualization plot. The default value, \code{0.2}, specifies that 20 \% of the plotting areas' widths are to be taken up by the bars for both trophic levels combined. The remaining 80 \% of the space will be taken up by the interactions (the edges).}

\item{Horizontal_Scale_Text_Shift_Constant}{a positive number specifying the relative horizontal distance (relative to the overall horizontal axis range) the scale bar text will be from the scale bar. The default value to this argument is \code{0.1}.}

\item{Vertical_Scale_Title_Shift_Constant}{a positive number specifying the relative vertical distance (relative to the overall vertical axis range) the scale bar title will be from the top of the scale bar. The default value to this argument is \code{0.1}.}

\item{Scale_Bar_Title_Font_Size}{the relative size of the scale bar titles. This argument must be a single positive number. The default value to this argument is \code{1}.}

\item{Scale_Bar_Text_Font_Size}{the relative size of the scale bar text. This argument must be a single positive number. The default value to this argument is \code{1}.}

\item{Overall_Plot_Title_Font_Size}{the relative size of the overall figure title. This argument must be a single positive number. The default value to this argument is \code{3.25}.}

\item{Trophic_Level_Plot_Title_Font_Size}{the relative size of the titles for the lower-trophic-level and higher-trophic-level title plots. This argument must be a single positive number. The default value to this argument is \code{2.5}.}

\item{Phenology_and_Bipartite_Plot_Title_Font_Size}{the relative size of the titles of the lower-trophic-level violin-like phenology plots, the plot containing lower-trophic-level species names, the bipartite interaction visualization plot, the plot containing the higher-trophic-level species names, and the higher-trophic-level violin-like phenology plots. This argument must be a single positive number. The default value to this argument is \code{1.5}.}

\item{Species_Name_Font_Size}{the relative size of the species names for the lower-trophic-level violin-like phenology plots and the higher-trophic-level violin-like phenology plots. This argument must be a single positive number. The default value to this argument is \code{1.5}.}

\item{Horizontal_Axis_Title_Font_Size}{the relative size of the horizontal axis titles for the lower-trophic-level violin-like phenology plots and the higher-trophic-level violin-like phenology plots. This argument must be a single positive number. The default value to this argument is \code{1}.}

\item{Horizontal_Axis_Dates_Font_Size}{the relative size of the dates on the horizontal axes of the lower-trophic-level violin-like phenology plots and higher-trophic-level violin-like phenology plots. This argument must be a single positive number. The default value to this argument is \code{1}.}

\item{Left_and_Right_Figure_Margins}{the relative distance between the left figure margin and the left edge of the lower-trophic-level violin-like phenology plots, and the relative distance between the right figure margin and the right edge of the higher-trophic-level violin-like phenology plots. This argument must be a single positive number. The default value to this argument is \code{9}.}

\item{Between_Title_Plots_and_Other_Plots_Margin}{the relative distance between the bottom of the title plots (which are on the top row of the figure) and the top of the plots on the bottom row of the figure. This argument must be a single positive number. The default value to this argument is \code{1}.}

\item{Bottom_Margin}{the relative distance between the edges of the plots on the bottom row of the figure and the lower figure margin. This argument must be a single positive number. The default value to this argument is \code{8}.}

\item{How_to_Treat_Missing_Values}{a character string that specifies how to handle missing values in the matrices. The three options for this argument are one of \code{"Remove the Rows"}, \code{"Remove the Columns"}, and \code{"Make Missing Values Zeros"}. The \code{"Make Missing Values Zeros"} option is not recommended because missing values do not necessarily mean that the species (the node) was not present. The default value, \code{"Remove the Columns"}, means that columns, not rows, which contain missing values will be removed (the lower-trophic-level species are preserved at the expense of the higher-trophic-level species).}
}
\value{
This function returns a figure with violin-like phenology plots showing the abundances of species from the lower trophic levels over time on the left side of the figure and violin-like phenology plots showing the abundances of species from the higher trophic levels over time on the right side of the figure. In the middle of these two is a bipartite interaction matrix visualization which shows how the two trophic levels interacted.
}
\description{
This function generates violin-like phenology plots that depict the relative abundances of lower-trophic-level and higher-trophic-level species over time. It also generates bipartite interaction matrix visualization plots which are positioned centrally between the two sets of phenology plots, with the lower-trophic-level information on the left and the higher-trophic-level information on the right.
}
\details{
Unlike the \link[DBM.functions]{Generating_Figures_Containing_a_Bipartite_Interaction_Matrix_Visualization_Plot_and_Corresponding_Higher_and_Lower_Trophic_Level_Phenology_Plots} function, which can only create one of these figures at a time, this function splits a balanced data set up by one or more factors and then arranges figures for each factor or combination of factors nicely. This function requires perfectly balanced experimental designs - for multifactorial experimental designs, all of the levels of each factor must be perfectly crossed with all of the levels of the other factors.

Unfortunately, this function can't handle unbalanced designs or missing values at this point!

\code{Generating_Figures_Containing_a_Bipartite_Interaction_Matrix_Visualization_Plot_and_Corresponding_Higher_and_Lower_Trophic_Level_Phenology_Plots_by_Factors} generates violin-like phenology plots that depict the relative abundances of lower-trophic-level and higher-trophic-level species over time. It also generates bipartite interaction matrix visualization plots which are positioned centrally between the two sets of phenology plots, with the lower-trophic-level information on the left and the higher-trophic-level information on the right.

In bipartite parlance, 'nodes' refer to species (from either trophic level) and 'edges' refer to interactions between these species. Bipartite networks are special because species interact only across, and not within, trophic levels.

This function was heavily inspired by a manuscript that contained a figure very similar to the one this function returns (Russo et al., 2013). It was also heavily inspired by my colleague Isaac Ativor's work in plant-pollinator interactions.

This function relies on the \link[DBM.functions]{Attempting_to_Diagonalize_a_Matrix_Based_on_Its_Full_Diagonal} function, the \link[DBM.functions]{Finding_the_Least_Common_Denominator_of_Imperfect_Fractions} function, the \link[DBM.functions]{Generating_Overall_Plot_Layout_Matrices} function, the \link[DBM.functions]{Making_Reasonable_Scale_Bars} function, and the \link[DBM.functions]{Making_Replicable_Layout_Matrices_for_Complicated_Figures} function, which are all also found in the \code{DBM.functions} package.
}
\examples{
# Make Up Some Data
Date <- as.Date(rep(c("2022-5-3", "2022-6-4", "2022-7-1", "2022-8-6", "2022-9-3"), each = 300))
Block <- rep(seq_len(5), each = 60, length.out = 1500)
Factor_1 <- rep(paste("Treatment", seq_len(2), sep = "_"), each = 30, length.out = 1500)
Factor_2 <- rep(paste("Treatment", seq_len(3), sep = "_"), each = 10, length.out = 1500)
Factor_3 <- rep(paste("Treatment", seq_len(2), sep = "_"), each = 5, length.out = 1500)
Lower_Trophic_Level_Species <- rep(paste("Species", letters[seq_len(5)], sep = "_"), length.out = 1500)
Data_Frame <- data.frame(Date = Date, Block = Block, Factor_1 = Factor_1, Factor_2 = Factor_2, Factor_3 = Factor_3, Lower_Trophic_Level_Species = Lower_Trophic_Level_Species)
set.seed(42)
Abundance_Matrix <- matrix(rpois(nrow(Data_Frame) * 6, rlnorm(1)), ncol = 6)
colnames(Abundance_Matrix) <- Higher_Trophic_Level_Species <- paste("Species", LETTERS[seq_len(6)], sep = "_")
Data_Frame <- cbind(Data_Frame, Abundance_Matrix)

# Generate Some Extra Data Frames to Ensure the Names on the Figures Are Suitable for Publication
Factor_Data_for_Plotting <- data.frame(Factor = c("Factor_1", "Factor_2", "Factor_3"), Factor_Name_for_Plotting = c("Factor 1", "Factor 2", "Factor 3"))
Factor_Level_Data_for_Plotting <- data.frame(Factor = c(rep("Factor_1", 2), rep("Factor_2", 3), rep("Factor_3", 2)), Level = paste("Treatment", c(1:2, 1:3, 1:2), sep = "_"), Level_Name_for_Plotting = paste("Treatment", c(1:2, 1:3, 1:2)))
Lower_Trophic_Level_Species_Names_Data_Frame <- data.frame(Lower_Trophic_Level_Species_Name = c("Species_a", "Species_b", "Species_c", "Species_d", "Species_e"), Names_for_Plotting = c("Species a", "Species b", "Species c", "Species d", "Species e"))
Higher_Trophic_Level_Species_Names_Data_Frame <- data.frame(Higher_Trophic_Level_Species_Name = c("Species_A", "Species_B", "Species_C", "Species_D", "Species_E", "Species_F"), Names_for_Plotting = c("Species A", "Species B", "Species C", "Species D", "Species E", "Species F"))

# Make the Figure
Generating_Figures_Containing_a_Bipartite_Interaction_Matrix_Visualization_Plot_and_Corresponding_Higher_and_Lower_Trophic_Level_Phenology_Plots_by_Factors(Data_Frame = Data_Frame, Higher_Trophic_Level_Species = c("Species_A", "Species_B", "Species_C", "Species_D", "Species_E", "Species_F"), Lower_Trophic_Level_Species = "Lower_Trophic_Level_Species", Date = "Date", Factors_to_Split_by = c("Factor_1", "Factor_2", "Factor_3"), Factor_Data_for_Plotting = Factor_Data_for_Plotting, Factor_Level_Data_for_Plotting = Factor_Level_Data_for_Plotting, Higher_Trophic_Level_Species_Names_Data_Frame = Higher_Trophic_Level_Species_Names_Data_Frame, Lower_Trophic_Level_Species_Names_Data_Frame = Lower_Trophic_Level_Species_Names_Data_Frame)

}
\references{
Russo, L., N. DeBarros, S. Yang, K. Shea, and D. Mortensen. 2013. Supporting crop pollinators with floral resources: network-based phenological matching. Ecol. Evol. 3:3125-3140.
}
\author{
David B. Moore (\email{davidblakneymoore@gmail.com})
}
