% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Generating_Figures_Containing_Lower_Trophic_Level_Phenology_Plots_a_Bipartite_Interaction_Matrix_Visualization_Plot_and_Upper_Trophic_Level_Phenology_Plots.R
\name{Generating_Figures_Containing_Lower_Trophic_Level_Phenology_Plots_a_Bipartite_Interaction_Matrix_Visualization_Plot_and_Upper_Trophic_Level_Phenology_Plots}
\alias{Generating_Figures_Containing_Lower_Trophic_Level_Phenology_Plots_a_Bipartite_Interaction_Matrix_Visualization_Plot_and_Upper_Trophic_Level_Phenology_Plots}
\title{Generating Figures Containing Lower-Trophic-Level Phenology Plots, a Bipartite Interaction Matrix Visualization Plot, and Upper-Trophic-Level Phenology Plots}
\usage{
Generating_Figures_Containing_Lower_Trophic_Level_Phenology_Plots_a_Bipartite_Interaction_Matrix_Visualization_Plot_and_Upper_Trophic_Level_Phenology_Plots(
  List_of_Bipartite_Interaction_Matrices_by_Date,
  Overall_Plot_Title = NULL,
  Lower_Trophic_Level_Plot_Title = "Lower Trophic Level",
  Higher_Trophic_Level_Plot_Title = "Higher Trophic Level",
  Prioritize_Interaction_Strength_Over_Alphabetization_in_Species_Layout = FALSE,
  Lower_Trophic_Level_Color = "darkgreen",
  Higher_Trophic_Level_Color = "darkred",
  Interaction_Color = "black",
  Include_Scale_Bars = TRUE,
  Scale_Bar_Factor = 0.25,
  Relative_Plot_Widths = c(11, 4, 10, 4, 11),
  Relative_Plot_Heights = c(rep(1, ifelse(is.null(Overall_Plot_Title), 1, 2)),
    ifelse(is.null(Overall_Plot_Title), 9, 8)),
  Gap_Size_Constant = 0.2,
  Relative_Bar_Width = 0.2,
  Horizontal_Scale_Text_Shift_Constant = 0.1,
  Vertical_Scale_Title_Shift_Constant = 0.1,
  Scale_Bar_Title_Font_Size = 1,
  Scale_Bar_Text_Font_Size = 1,
  Overall_Plot_Title_Font_Size = 3.25,
  Trophic_Level_Plot_Title_Font_Size = 2.5,
  Phenology_and_Bipartite_Plot_Title_Font_Size = 1.5,
  Species_Name_Font_Size = 1.5,
  Horizontal_Axis_Title_Font_Size = 1,
  Horizontal_Axis_Dates_Font_Size = 1,
  Left_and_Right_Figure_Margins = 9,
  Between_Title_Plots_and_Other_Plots_Margin = 1,
  Bottom_Margin = 8
)
}
\arguments{
\item{List_of_Bipartite_Interaction_Matrices_by_Date}{a list of matrices. Each matrix contains quantitative information about the interactions between lower-trophic-level species and higher-trophic-level species. Each row must represent a lower-trophic-level species and each column must represent a higher-trophic-level species; similarly, the row names must be the names of the lower-trophic-level species and the column names must be the names of the higher-trophic-level species. Finally, the names of each matrix must be the date the observations were made, with the dates following the typical \code{"Date"} format that R uses (\code{"\%Y-\%m-\%d"}). No missing values are allowed in the matrices.}

\item{Overall_Plot_Title}{the overall figure title. The default value, \code{NULL}, means that an overall figure title will be omitted.}

\item{Lower_Trophic_Level_Plot_Title}{a name for the lower trophic level, such as \code{"Plants"} or \code{"Prey"}. The default value, \code{"Lower Trophic Level"}, is a generic title that can be used in any figure.}

\item{Higher_Trophic_Level_Plot_Title}{a name for the higher trophic level, such as \code{"Pollinators"} or \code{"Predators"}. The default value, \code{"Higher Trophic Level"}, is a generic title that can be used in any figure.}

\item{Prioritize_Interaction_Strength_Over_Alphabetization_in_Species_Layout}{whether species should be arranged such that species from different trophic levels that interact more frequently are closer together and species that interact less frequently are farther apart on the bipartite interaction visualization plot. If putting frequently interacting species closer together and infrequently interacting species farther apart is prioritized, the species will probably not be arranged alphabetically in the resulting figure. The default value, \code{FALSE}, causes the species from both trophic levels to be arranged alphabetically.}

\item{Lower_Trophic_Level_Color}{the color of the violin-like phenology plots and the bars (from the bipartite interaction visualization plot) for the lower-trophic-level species. The default value to this argument is \code{"darkgreen"}.}

\item{Higher_Trophic_Level_Color}{the color of the violin-like phenology plots and the bars (from the bipartite interaction visualization plot) for the higher-trophic-level species. The default value to this argument is \code{"darkred"}.}

\item{Interaction_Color}{the color of the edges (from the bipartite interaction visualization plot) for the interactions between the lower-trophic-level species and the higher-trophic-level species. The default value to this argument is \code{"black"}.}

\item{Include_Scale_Bars}{whether vertical scale bars should be included on the left and right sides of the figure. The default value, \code{TRUE}, causes scale bars to be included.}

\item{Scale_Bar_Factor}{a number between 0 and 1 (exclusive) that specifies approximately how tall the scale bar will be relative to the vertical axis range. The default value for this argument, \code{0.25}, makes it so that the scale bar will take up as close to 25 \% of the vertical axis' range as possible.}

\item{Relative_Plot_Widths}{the relative widths of the lower-trophic-level violin-like phenology plots, the plot containing lower-trophic-level species names, the bipartite interaction visualization plot, the plot containing the higher-trophic-level species names, and the higher-trophic-level violin-like phenology plots. This argument must be a numeric vector containing 5 positive numbers, and it must be the same forward and backward to ensure that the resulting figure is symmetric. The default value to this argument is \code{c(11, 4, 10, 4, 11)}.}

\item{Relative_Plot_Heights}{the relative heights of the plots containing the titles and all the other plots. This argument must be a numeric vector contain two (if an overall figure title is not to be included) or three (if an overall figure title is to be included) positive numbers. The default value to this argument, \code{c(rep(1, ifelse(is.null(Overall_Plot_Title), 1, 2)), ifelse(is.null(Overall_Plot_Title), 9, 8))}, means that when there is no overall figure title, the heights will be \code{c(1, 9)}, and when there is an overall figure title, the heights will be \code{c(1, 8, 8)}.}

\item{Gap_Size_Constant}{a nonnegative number that specifies how much space is allocated between bars vertically (within trophic levels) in the bipartite interaction visualization plot, where each bar represents a species. Using higher numbers for this argument will lead to more space between species' bars. The default value, \code{0.2}, specifies that 20 \% of the cumulative bar heights (without any space) will be divided up equally between the bars in this plot.}

\item{Relative_Bar_Width}{a number between \code{0} and \code{1} (inclusive) that specifies how wide the bars representing species are in the bipartite interaction visualization plot. The default value, \code{0.2}, specifies that 20 \% of the plotting areas' widths are to be taken up by the bars for both trophic levels combined. The remaining 80 \% of the space will be taken up by the interactions (the edges).}

\item{Horizontal_Scale_Text_Shift_Constant}{a positive number specifying the relative horizontal distance (relative to the overall horizontal axis range) the scale bar text will be from the scale bar. The default value to this argument is \code{0.1}.}

\item{Vertical_Scale_Title_Shift_Constant}{a positive number specifying the relative vertical distance (relative to the overall vertical axis range) the scale bar title will be from the top of the scale bar. The default value to this argument is \code{0.1}.}

\item{Scale_Bar_Title_Font_Size}{the relative size of the scale bar titles. This argument must be a single positive number. The default value to this argument is \code{1}.}

\item{Scale_Bar_Text_Font_Size}{the relative size of the scale bar text. This argument must be a single positive number. The default value to this argument is \code{1}.}

\item{Overall_Plot_Title_Font_Size}{the relative size of the overall figure title. This argument must be a single positive number. The default value to this argument is \code{3.25}.}

\item{Trophic_Level_Plot_Title_Font_Size}{the relative size of the titles for the lower-trophic-level and higher-trophic-level title plots. This argument must be a single positive number. The default value to this argument is \code{2.5}.}

\item{Phenology_and_Bipartite_Plot_Title_Font_Size}{the relative size of the titles of the lower-trophic-level violin-like phenology plots, the plot containing lower-trophic-level species names, the bipartite interaction visualization plot, the plot containing the higher-trophic-level species names, and the higher-trophic-level violin-like phenology plots. This argument must be a single positive number. The default value to this argument is \code{1.5}.}

\item{Species_Name_Font_Size}{the relative size of the species names for the lower-trophic-level violin-like phenology plots and the higher-trophic-level violin-like phenology plots. This argument must be a single positive number. The default value to this argument is \code{1.5}.}

\item{Horizontal_Axis_Title_Font_Size}{the relative size of the horizontal axis titles for the lower-trophic-level violin-like phenology plots and the higher-trophic-level violin-like phenology plots. This argument must be a single positive number. The default value to this argument is \code{1}.}

\item{Horizontal_Axis_Dates_Font_Size}{the relative size of the dates on the horizontal axes of the lower-trophic-level violin-like phenology plots and higher-trophic-level violin-like phenology plots. This argument must be a single positive number. The default value to this argument is \code{1}.}

\item{Left_and_Right_Figure_Margins}{the relative distance between the left figure margin and the left edge of the lower-trophic-level violin-like phenology plots, and the relative distance between the right figure margin and the right edge of the higher-trophic-level violin-like phenology plots. This argument must be a single positive number. The default value to this argument is \code{9}.}

\item{Between_Title_Plots_and_Other_Plots_Margin}{the relative distance between the bottom of the title plots (which are on the top row of the figure) and the top of the plots on the bottom row of the figure. This argument must be a single positive number. The default value to this argument is \code{1}.}

\item{Bottom_Margin}{the relative distance between the edges of the plots on the bottom row of the figure and the lower figure margin. This argument must be a single positive number. The default value to this argument is \code{8}.}
}
\value{
This function returns a figure with violin-like phenology plots showing the abundances of species from the lower trophic levels over time on the left side of the figure and violin-like phenology plots showing the abundances of species from the higher trophic levels over time on the right side of the figure. In the middle of these two is a bipartite interaction matrix visualization which shows how the two trophic levels interacted.
}
\description{
This function generates violin-like phenology plots that depict the relative abundances of lower-trophic-level and higher-trophic-level species over time. It also generates bipartite interaction matrix visualization plots which are positioned centrally between the two sets of phenology plots.
}
\details{
\code{Generating_Figures_Containing_Lower_Trophic_Level_Phenology_Plots_a_Bipartite_Interaction_Matrix_Visualization_Plot_and_Upper_Trophic_Level_Phenology_Plots} generates violin-like phenology plots that depict the relative abundances of lower-trophic-level and higher-trophic-level species over time. It also generates bipartite interaction matrix visualization plots which are positioned centrally between the two sets of phenology plots.

In bipartite parlance, 'nodes' refer to species (from either trophic level) and 'edges' refer to interactions between these species. Bipartite networks are special because species interact only across, and not within, trophic levels.

This function was heavily inspired by a manuscript that contained a figure very similar to the one this function returns (Russo et al., 2013). It was also heavily inspired by my colleague Isaac Ativor's work in plant-pollinator interactions.
}
\examples{
# Generate Made-Up Data
set.seed(21)
List_of_Bipartite_Interaction_Matrices_by_Date_2 <- list("2020-04-05" = matrix(round(rchisq(28, 6)), ncol = 7), "2020-05-10" = matrix(round(rf(28, 5, 2)), ncol = 7), "2020-06-09" = matrix(round(rlnorm(28)), ncol = 7), "2020-07-13" = matrix(round(rexp(28)), ncol = 7))
List_of_Bipartite_Interaction_Matrices_by_Date_2 <- lapply(List_of_Bipartite_Interaction_Matrices_by_Date_2, function (x) {
rownames(x) <- paste("Species", letters[seq_len(nrow(x))])
  colnames(x) <- paste("Species", LETTERS[seq_len(ncol(x))])
  x
})

# Example Figure 1
Generating_Figures_Containing_Lower_Trophic_Level_Phenology_Plots_a_Bipartite_Interaction_Matrix_Visualization_Plot_and_Upper_Trophic_Level_Phenology_Plots(List_of_Bipartite_Interaction_Matrices_by_Date_2, Scale_Bar_Factor = 0.25, Include_Scale_Bars = FALSE)

# Example Figure 2
Generating_Figures_Containing_Lower_Trophic_Level_Phenology_Plots_a_Bipartite_Interaction_Matrix_Visualization_Plot_and_Upper_Trophic_Level_Phenology_Plots(List_of_Bipartite_Interaction_Matrices_by_Date_2, Scale_Bar_Factor = 0.25, Prioritize_Interaction_Strength_Over_Alphabetization_in_Species_Layout = TRUE, Overall_Plot_Title = "Data From All Treatments Combined", Lower_Trophic_Level_Plot_Title = "Plants", Higher_Trophic_Level_Plot_Title = "Pollinators")

}
\references{
Russo, L., N. DeBarros, S. Yang, K. Shea, and D. Mortensen. 2013. Supporting crop pollinators with floral resources: network-based phenological matching. Ecol. Evol. 3:3125-3140.
}
\author{
David B. Moore (\email{davidblakneymoore@gmail.com})
}
