% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generating_Permutations.R
\name{Generating_Permutations}
\alias{Generating_Permutations}
\title{Generating Permutations}
\usage{
Generating_Permutations(Number_of_Items)
}
\arguments{
\item{Number_of_Items}{a positive integer for which permutations will be calculated.}
}
\value{
This function returns a list of all the possible permutations of a given number of items.
}
\description{
This function generates all possible permutations for a given number of items.
}
\details{
\code{Generating_Permutations} generates all possible permutations for a given number of items.
}
\examples{
Generating_Permutations(0)
Generating_Permutations(1)
Generating_Permutations(4)

}
\author{
David B. Moore (\email{davidblakneymoore@gmail.com})
}
