% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Laying_Out_Plots_Across_Multiple_Pages_or_Slides.R
\name{Laying_Out_Plots_Across_Multiple_Pages_or_Slides}
\alias{Laying_Out_Plots_Across_Multiple_Pages_or_Slides}
\title{Laying Out Plots Across Multiple Pages or Slides}
\usage{
Laying_Out_Plots_Across_Multiple_Pages_or_Slides(
  Number_of_Plots,
  Number_of_Rows,
  Number_of_Columns,
  Add_a_Row_for_a_Figure_Title = T,
  Add_a_Row_for_a_Figure_Legend = T,
  Should_Blank_Space_Be_Left_at_the_Bottom_of_the_Last_Figure = T
)
}
\arguments{
\item{Number_of_Plots}{the number of plots to be arranged across multiple figures (across multiple pages or slides).}

\item{Number_of_Rows}{the number of rows of plots within each figure.}

\item{Number_of_Columns}{the number of columns of plots within each figure.}

\item{Add_a_Row_for_a_Figure_Title}{a logical argument indicating if a row should be added at the top of each layout matrix for a title plot. The default to this argument, \code{TRUE}, adds a row at the top of each layout matrix for a title plot.}

\item{Add_a_Row_for_a_Figure_Legend}{a logical argument indicating if a row should be added at the bottom of each layout matrix for a legend plot. The default to this argument, \code{TRUE}, adds a row at the bottom of each layout matrix for a legend plot.}

\item{Should_Blank_Space_Be_Left_at_the_Bottom_of_the_Last_Figure}{a logical argument indicating if blank space should be left at the bottom of the last layout matrix if the last layout matrix contains fewer rows than the previous layout matrix or layout matrices to ensure that the last layout matrix is the same size as the previous layout matrix or layout matrices. When creating figures within R, it's best if this argument is \code{TRUE}, whereas when creating figures that will be stored to the hard drive, it's best if this argument is \code{FALSE} (see the examples for more details). The default to this argument, \code{TRUE}, adds blank space if necessary to ensure that all layout matrices are the exact same size (which, again, is particularly nice when printing figures in R).}
}
\value{
This function returns a list of plot layout matrices (\code{List_of_Layout_Matrices}) where each list element corresponds with a plot layout matrix for a new figure (a new page of a manuscript or a new slide of a presentation). This function also returns a vector (\code{Indices}) that should be used to assign plots to figures (see the examples for more details).
}
\description{
This function generates plot layout matrices that allow the user to break up extensive numbers of plots neatly across multiple figures across multiple pages of a manuscript or multiple slides in a presentation.
}
\details{
\code{Laying_Out_Plots_Across_Multiple_Pages_or_Slides} generates plot layout matrices that allow the user to break up extensive numbers of plots neatly across multiple figures across multiple pages of a manuscript or multiple slides in a presentation.
}
\examples{
Data_Frame <- as.data.frame(matrix(rnorm(384), ncol = 32))
colnames(Data_Frame) <- paste("Response_Variable", seq_len(ncol(Data_Frame)), sep = "_")
Data_Frame$Group <- rep(LETTERS[1:4], each = 3)
Data_Frame <- Data_Frame[, c(which(colnames(Data_Frame) == "Group"), which(colnames(Data_Frame) != "Group"))]
Relative_Title_Plot_Height <- 1
Relative_Boxplot_Plot_Height <- 4

# Example 1: Printing Figures in R
Output <- Laying_Out_Plots_Across_Multiple_Pages_or_Slides(Number_of_Plots = ncol(Data_Frame[, grep("Response_Variable", colnames(Data_Frame))]), Number_of_Rows = 4, Number_of_Columns = 3, Add_a_Row_for_a_Figure_Legend = F)
lapply(seq_along(Output$List_of_Layout_Matrices), function (i) {
  Data_Subset <- Data_Frame[, c(which(colnames(Data_Frame) == "Group"), grep("Response_Variable", colnames(Data_Frame))[which(Output$Indices == i)])]
  layout(Output$List_of_Layout_Matrices[[i]], heights = c(Relative_Title_Plot_Height, rep(Relative_Boxplot_Plot_Height, (nrow(Output$List_of_Layout_Matrices[[i]]) - 1))))
  par(mar = c(1, 1, 1, 1))
  plot(0, type = "n", axes = F, xlab = "", ylab = "", xlim = c(-1, 1), ylim = c(-1, 1))
  if (i == 1) {
    text(0, 0, expression(paste(bold("Boxplots"))), cex = 2.5)
  } else if (i > 1) {
    text(0, 0, expression(paste(bold("Boxplots, continued"))), cex = 2.5)
  }
  par(mar = c(5, 5, 4, 2))
  lapply(seq_along(grep("Response_Variable", colnames(Data_Subset))), function (j) {
    boxplot(Data_Subset[, grep("Response_Variable", colnames(Data_Subset))[j]] ~ Data_Subset$Group, main = gsub("_", " ", colnames(Data_Subset)[grep("Response_Variable", colnames(Data_Subset))[j]]), xlab = "Group", ylab = "Response", cex.axis = 1.5, cex.lab = 1.5, cex.main = 2)
  })
})

# Example 2: Saving Figures to the Hard Drive
# Caution - this chunk of code will actually store figures in your working directory on your hard drive named 'Boxplot Figure 1', 'Boxplot Figure 2', and 'Boxplot Figure 3'!
Output <- Laying_Out_Plots_Across_Multiple_Pages_or_Slides(Number_of_Plots = ncol(Data_Frame[, grep("Response_Variable", colnames(Data_Frame))]), Number_of_Rows = 4, Number_of_Columns = 3, Add_a_Row_for_a_Figure_Legend = F, Should_Blank_Space_Be_Left_at_the_Bottom_of_the_Last_Figure = F)
Horizontal_Figure_Stretching_Constant <- 75
Vertical_Figure_Stretching_Constant <- 75
lapply(seq_along(Output$List_of_Layout_Matrices), function (i) {
  Data_Subset <- Data_Frame[, c(which(colnames(Data_Frame) == "Group"), grep("Response_Variable", colnames(Data_Frame))[which(Output$Indices == i)])]
  jpeg(paste0("Boxplot Figure ", i, ".jpeg"), height = ((Vertical_Figure_Stretching_Constant * (nrow(Output$List_of_Layout_Matrices[[i]]) - 1) * Relative_Boxplot_Plot_Height) + (Vertical_Figure_Stretching_Constant * Relative_Title_Plot_Height)), width = (Horizontal_Figure_Stretching_Constant * ncol(Output$List_of_Layout_Matrices[[i]]) * Relative_Boxplot_Plot_Height / 2))
  layout(Output$List_of_Layout_Matrices[[i]], heights = c(Relative_Title_Plot_Height, rep(Relative_Boxplot_Plot_Height, (nrow(Output$List_of_Layout_Matrices[[i]]) - 1))))
  par(mar = c(1, 1, 1, 1))
  plot(0, type = "n", axes = F, xlab = "", ylab = "", xlim = c(-1, 1), ylim = c(-1, 1))
  if (i == 1) {
    text(0, 0, expression(paste(bold("Boxplots"))), cex = 2.5)
  } else if (i > 1) {
    text(0, 0, expression(paste(bold("Boxplots, continued"))), cex = 2.5)
  }
  par(mar = c(5, 5, 4, 2))
  lapply(seq_along(grep("Response_Variable", colnames(Data_Subset))), function (j) {
    boxplot(Data_Subset[, grep("Response_Variable", colnames(Data_Subset))[j]] ~ Data_Subset$Group, main = gsub("_", " ", colnames(Data_Subset)[grep("Response_Variable", colnames(Data_Subset))[j]]), xlab = "Group", ylab = "Response", cex.axis = 1.5, cex.lab = 1.5, cex.main = 2)
  })
  dev.off()
})

}
\author{
David B. Moore (\email{davidblakneymoore@gmail.com})
}
