% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/data-Made_Up_Data_for_the_Aligning_Values_Across_Multiple_Vertical_Axes_Function.R
\docType{data}
\name{Made_Up_Data_for_the_Aligning_Values_Across_Multiple_Vertical_Axes_Function}
\alias{Made_Up_Data_for_the_Aligning_Values_Across_Multiple_Vertical_Axes_Function}
\title{Made-up Data for the 'DBM.functions::Aligning_Values_Across_Multiple_Vertical_Axes' Function}
\format{
\code{Made_Up_Data_for_the_Aligning_Values_Across_Multiple_Vertical_Axes_Function} is a \code{data.frame} with 100 rows and 4 columns.
\describe{
\item{Index}{An indexing variable}
\item{Variable_1, Variable_2, Variable_3}{Variables (each of which is measured in different units) that are to be aligned across vertical axes on a graph}
}
}
\usage{
Made_Up_Data_for_the_Aligning_Values_Across_Multiple_Vertical_Axes_Function
}
\description{
This data frame contains made-up data to use in the example in the documentation of the 'DBM.functions::Aligning_Values_Across_Multiple_Vertical_Axes' function.
}
\examples{
Made_Up_Data_for_the_Aligning_Values_Across_Multiple_Vertical_Axes_Function
}
\author{
David B. Moore (\email{davidblakneymoore@gmail.com})
}
\keyword{datasets}
