% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/data-Made_Up_Data_for_the_Comparing_Multiple_Independent_Correlation_Coefficients_Function.R
\docType{data}
\name{Made_Up_Data_for_the_Comparing_Multiple_Independent_Correlation_Coefficients_Function}
\alias{Made_Up_Data_for_the_Comparing_Multiple_Independent_Correlation_Coefficients_Function}
\title{Made-up Data for the 'Comparing_Multiple_Independent_Correlation_Coefficients' Function}
\format{
\code{Made_Up_Data_for_the_Comparing_Multiple_Independent_Correlation_Coefficients_Function} is a \code{data.frame} with 5 rows and 3 columns.
\describe{
\item{Name}{The name of the correlation}
\item{Coefficient_of_Correlation}{The correlation coefficient}
\item{Sample_Size}{The number of observations that went into calculating the correlation coefficient}
}
}
\usage{
Made_Up_Data_for_the_Comparing_Multiple_Independent_Correlation_Coefficients_Function
}
\description{
This data frame contains made-up data to use in the example in the documentation of the 'Comparing_Multiple_Independent_Correlation_Coefficients' function.
}
\examples{
Made_Up_Data_for_the_Comparing_Multiple_Independent_Correlation_Coefficients_Function
}
\author{
David B. Moore (\email{davidblakneymoore@gmail.com})
}
\keyword{datasets}
