% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Optimally_Assigning_Experimental_Units_to_Treatment_Groups_for_Completely_Randomized_Designs.R
\name{Optimally_Assigning_Experimental_Units_to_Treatment_Groups_for_Completely_Randomized_Designs}
\alias{Optimally_Assigning_Experimental_Units_to_Treatment_Groups_for_Completely_Randomized_Designs}
\title{Optimally Assigning Experimental Units to Treatment Groups for Completely Randomized Designs}
\usage{
Optimally_Assigning_Experimental_Units_to_Treatment_Groups_for_Completely_Randomized_Designs(
  ...,
  Identifiers,
  Data_Frame = NULL,
  Number_of_Treatment_Groups,
  Number_of_Experimental_Units_in_Each_Treatment_Group,
  Variable_Weights = rep(1, ifelse(missing(Data_Frame), length(list(...)),
    ncol(Data_Frame[, c(which(colnames(Data_Frame) \%in\% sapply(match.call(expand.dots =
    FALSE)$..., deparse)))]))),
  Moment_Weighing_Method = c("Exponential (2 Base)", "Exponential (Golden Ratio Base)",
    "Ordinal", "Exponential (Euler's Number Base)", "None", "Custom Weights"),
  Custom_Moment_Weights = NULL,
  Number_of_Combinations_to_Report = ifelse(factorial(ifelse(missing(Data_Frame),
    unique(sapply(list(...), length)),
    nrow(Data_Frame)))/(factorial(Number_of_Treatment_Groups) * Reduce(`*`,
    sapply(seq_len(Number_of_Treatment_Groups), function(x) {
    
    factorial(Number_of_Experimental_Units_in_Each_Treatment_Group)
 }))) <= 3,
    factorial(ifelse(missing(Data_Frame), unique(sapply(list(...), length)),
    nrow(Data_Frame)))/(factorial(Number_of_Treatment_Groups) * Reduce(`*`,
    sapply(seq_len(Number_of_Treatment_Groups), function(x) {
    
    factorial(Number_of_Experimental_Units_in_Each_Treatment_Group)
 }))),
    ifelse((floor((factorial(ifelse(missing(Data_Frame), unique(sapply(list(...),
    length)), nrow(Data_Frame)))/(factorial(Number_of_Treatment_Groups) * Reduce(`*`,
    sapply(seq_len(Number_of_Treatment_Groups), function(x) {
    
    factorial(Number_of_Experimental_Units_in_Each_Treatment_Group)
 })))) * 0.01) <= 3)
    & (factorial(ifelse(missing(Data_Frame), unique(sapply(list(...), length)),
    nrow(Data_Frame)))/(factorial(Number_of_Treatment_Groups) * Reduce(`*`,
    sapply(seq_len(Number_of_Treatment_Groups), function(x) {
    
    factorial(Number_of_Experimental_Units_in_Each_Treatment_Group)
 }))) > 3), 3,
    ifelse(floor((factorial(ifelse(missing(Data_Frame), unique(sapply(list(...),
    length)), nrow(Data_Frame)))/(factorial(Number_of_Treatment_Groups) * Reduce(`*`,
    sapply(seq_len(Number_of_Treatment_Groups), function(x) {
    
    factorial(Number_of_Experimental_Units_in_Each_Treatment_Group)
 })))) * 0.01) > 3,
    floor((factorial(ifelse(missing(Data_Frame), unique(sapply(list(...), length)),
    nrow(Data_Frame)))/(factorial(Number_of_Treatment_Groups) * Reduce(`*`,
    sapply(seq_len(Number_of_Treatment_Groups), function(x) {
    
    factorial(Number_of_Experimental_Units_in_Each_Treatment_Group)
 })))) * 0.01))))
)
}
\arguments{
\item{...}{the variables you measured on your experimental units that you wish to keep as consistent as possible across treatment groups.}

\item{Identifiers}{the names of the experimental units.}

\item{Data_Frame}{an optional argument to provide if all the \code{...} arguments and the \code{Identifiers} argument all come from the same data frame and to prevent typing the data frame name multiple times.}

\item{Number_of_Treatment_Groups}{the number of treatment groups in the study or experiment.}

\item{Number_of_Experimental_Units_in_Each_Treatment_Group}{the number of experimental units to be assigned to each treatment group in the study or experiment. This function only handles balanced designs where each treatment group contains the same number of experimental units.}

\item{Variable_Weights}{the relative weights (importances) of each of the \code{...} arguments (variables) in the calculations. Variables that are weighed more heavily will be held more consistent across treatment groups than variables that are weighed less heavily. By default, each variable is weighed equally.}

\item{Moment_Weighing_Method}{the method of weighing mathematical moments. Almost all of the built-in methods weigh moments sequentially, with lower-order moments being weighed more heavily than higher-order moments are weighed. This function calculates moments up to the order of the number of experimental units in each treatment group (as specified in the \code{Number_of_Experimental_Units_in_Each_Treatment_Group} argument). The default method weighs each moment twice as heavily as the subsequent moment, with lower-order moments being weighed the most heavily and higher-order moments being weighed the least heavily.}

\item{Custom_Moment_Weights}{the weights of each of the mathematical moments (in order from the lowest-order moment to the highest-order moment) if and only if \code{"Custom Weights"} is specified for the \code{Moment_Weighing_Method} argument.}

\item{Number_of_Combinations_to_Report}{the number of grouping assignments to report. Grouping assignments are always reported such that the best (most optimal) assignments are reported first. By default, for most applications, the top 1 \% of grouping assignments are reported.}
}
\value{
This function returns optimal treatment group assignments for completely randomized designs.
}
\description{
This function may be used to assign experimental units to treatment groups for completely randomized designs when you wish to ensure that treatment groups are as balanced as possible for a particular set of experimental units' variables.
}
\details{
\code{Optimally_Assigning_Experimental_Units_to_Treatment_Groups_for_Completely_Randomized_Designs} may be used when setting out a completely-randomized-design study or experiment to optimize the assignments of experimental units to treatment groups. Although randomization is typically used to assign treatments to experimental units, sometimes, randomization may unintentionally create treatment groups that are not balanced. For example, perhaps after randomizing, certain treatment groups end up being comprised of experimental units that are bigger or heavier or hotter or shadier or wetter than the experimental units in the other treatment groups are. This undesirable effect may be avoided by using this function, which ensures that means and possibly other higher-order mathematical moments (such as variances, skewnesses, and kurtoses) for a particular set of experimental units' variables you've already measured are as similar as possible across treatment groups.

Please note that randomization is still a very good technique. If you suspect that you did not measure a suite of experimental unit variables that may influence the response variables in your study or experiment, randomization may be a better method than this optimization method.
}
\examples{
set.seed(13)
Data_Frame <- data.frame(Tree_Identifier = paste("Tree", seq_len(13), sep = "_"), Diameter_at_Breast_Height = round(abs(rnorm(13, 30, 2.5)), 1), Height = round(abs(rnorm(13, 30, 2.5)), 1), Crown_Class = sample(c("Dominant", "Codominant", "Intermediate", "Overtopped"), 13, replace = TRUE))
(Optimal_Group_Assignments <- Optimally_Assigning_Experimental_Units_to_Treatment_Groups_for_Completely_Randomized_Designs(Diameter_at_Breast_Height, Height, Crown_Class, Identifiers = Tree_Identifier, Data_Frame = Data_Frame, Number_of_Combinations_to_Report = 3, Number_of_Treatment_Groups = 3, Number_of_Experimental_Units_in_Each_Treatment_Group = 4))

}
\author{
David B. Moore (\email{davidblakneymoore@gmail.com})
}
