% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ordinalizing_Integers.R
\name{Ordinalizing_Integers}
\alias{Ordinalizing_Integers}
\title{Ordinalizing Integers}
\usage{
Ordinalizing_Integers(Integer_Object)
}
\arguments{
\item{Integer_Object}{an R object containing integers.}
}
\value{
This function returns integers followed by their appropriate ordinal suffix.
}
\description{
This function takes integers and returns character strings comprised of the integer followed by its appropriate ordinal suffix.
}
\details{
\code{Ordinalizing_Integers} takes integers and returns character strings comprised of the integer followed by its appropriate ordinal suffix.
}
\examples{
Ordinalizing_Integers(1:25)

}
\author{
David B. Moore (\email{davidblakneymoore@gmail.com})
}
