% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Regression_for_Repeated_Measures_Designs_When_Predictor_Variables_Are_Bins_of_Numbers_and_Bins_Are_Continuous.R
\name{Regression_for_Repeated_Measures_Designs_When_Predictor_Variables_Are_Bins_of_Numbers_and_Bins_Are_Continuous}
\alias{Regression_for_Repeated_Measures_Designs_When_Predictor_Variables_Are_Bins_of_Numbers_and_Bins_Are_Continuous}
\title{Regression for Repeated-Measures Designs When Predictor Variables Are Bins of Numbers and Bins Are Continuous}
\usage{
Regression_for_Repeated_Measures_Designs_When_Predictor_Variables_Are_Bins_of_Numbers_and_Bins_Are_Continuous(
  Treatment,
  Replicate,
  Predictor_Variable,
  Predictor_Variable_Bin_Lower_Limits,
  Predictor_Variable_Bin_Upper_Limits,
  Response_Variable,
  Beta_Distribution_Alpha_Parameter_Values,
  Beta_Distribution_Beta_Parameter_Values,
  Data_Frame,
  Parameters,
  Formula,
  Color_Data,
  Color_Opacity = 0.5,
  Number_of_Simulations_per_Experimental_Unit = 1000,
  Number_of_Histogram_Bins = 100,
  Moments_to_Report = 4,
  Make_Figure_1 = T,
  Make_Figure_2 = T,
  Make_Figure_3 = T,
  Make_Figure_4 = T,
  Working_Directory = getwd(),
  First_Figure_Name = "Figure 1",
  Second_Figure_Name = "Figure 2",
  Third_Figure_Name = "Figure 3",
  Fourth_Figure_Name = "Figure 4",
  Show_Moments_in_Figure_1 = T,
  Show_Overlapping_Areas_in_Figure_2 = T,
  Show_Moments_in_Figure_3 = T,
  Show_Moments_in_Figure_4 = T,
  Moment_Rounding_Constant = 3,
  Overlapping_Area_Rounding_Constant = 3,
  Number_of_Bins_in_the_Figure_4_Histograms = 10,
  Maximum_Number_of_Iterations = 1000,
  Relative_Figure_1_Parameter_Plot_Height = 4,
  Relative_Figure_1_Treatment_Plot_Height = 3,
  Relative_Figure_1_Replicate_Plot_Height = 2,
  Relative_Figure_1_Individual_Plot_Height = 10,
  Relative_Figure_1_Legend_Plot_Height = 4,
  Relative_Figure_1_Individual_Plot_Width = 5,
  Relative_Figure_1_Gap_Plot_Width = 1,
  Relative_Figure_1_Parameter_Spacing_Line_Plot_Height = 1,
  Figure_1_Figure_Height_Constant = 30,
  Figure_1_Figure_Width_Constant = 45,
  Figure_1_Axis_Labels_and_Axis_Titles_Shifting_Constant = 0,
  Figure_1_Parameter_Title_Text_Size = 4,
  Figure_1_Treatment_Title_Text_Size = 3,
  Figure_1_Replicate_Title_Text_Size = 2,
  Figure_1_Axis_Titles_Text_Size = 2,
  Figure_1_Axis_Labels_Text_Size = 2,
  Figure_1_Individual_Plot_Bottom_Margin = 5,
  Figure_1_Individual_Plot_Left_Margin = 5,
  Figure_1_Individual_Plot_Top_Margin = 1,
  Figure_1_Individual_Plot_Right_Margin = 1,
  Figure_1_Moment_Text_Size = 2,
  Figure_1_Legend_Text_Size = 3,
  Figure_1_Panel_Separation_Line_Width = 2,
  Relative_Figure_2_Parameter_Plot_Height = 4,
  Relative_Figure_2_Treatment_Plot_Height = 3,
  Relative_Figure_2_Replicate_Plot_Height = 2,
  Relative_Figure_2_Individual_Plot_Dimension = 10,
  Relative_Figure_2_Gap_Plot_Dimension = 2,
  Relative_Figure_2_Legend_Plot_Height = 8,
  Relative_Figure_2_Parameter_Spacing_Line_Plot_Height = 1,
  Figure_2_Figure_Height_Constant = 75,
  Figure_2_Figure_Width_Constant = 100,
  Figure_2_Axis_Labels_and_Axis_Titles_Shifting_Constant = 1,
  Figure_2_Parameter_Title_Text_Size = 10,
  Figure_2_Treatment_Title_Text_Size = 8,
  Figure_2_Replicate_Title_Text_Size = 6,
  Figure_2_Axis_Titles_Text_Size = 2.5,
  Figure_2_Axis_Labels_Text_Size = 2.5,
  Figure_2_Individual_Plot_Bottom_Margin = 10,
  Figure_2_Individual_Plot_Left_Margin = 10,
  Figure_2_Individual_Plot_Top_Margin = 1,
  Figure_2_Individual_Plot_Right_Margin = 1,
  Figure_2_Overlapping_Area_Text_Size = 4,
  Figure_2_Legend_Text_Size = 3.5,
  Figure_2_Panel_Separation_Line_Width = 2,
  Relative_Figure_3_Parameter_Plot_Height = 4,
  Relative_Figure_3_Treatment_Plot_Dimension = 3,
  Relative_Figure_3_Replicate_Plot_Dimension = 2,
  Relative_Figure_3_Individual_Plot_Dimension = 10,
  Relative_Figure_3_Gap_Plot_Dimension = 2,
  Relative_Figure_3_Legend_Plot_Height = 4,
  Relative_Figure_3_Parameter_Spacing_Line_Plot_Height = 1,
  Figure_3_Figure_Height_Constant = 60,
  Figure_3_Figure_Width_Constant = 75,
  Figure_3_Axis_Labels_and_Axis_Titles_Shifting_Constant = 1,
  Figure_3_Parameter_Title_Text_Size = 10,
  Figure_3_Treatment_Title_Text_Size = 8,
  Figure_3_Replicate_Title_Text_Size = 6,
  Figure_3_Axis_Titles_Text_Size = 4,
  Figure_3_Axis_Labels_Text_Size = 4,
  Figure_3_Individual_Plot_Bottom_Margin = 10,
  Figure_3_Individual_Plot_Left_Margin = 10,
  Figure_3_Individual_Plot_Top_Margin = 1,
  Figure_3_Individual_Plot_Right_Margin = 1,
  Figure_3_Moment_Text_Size = 4,
  Figure_3_Legend_Text_Size = 3.5,
  Figure_3_Panel_Separation_Line_Width = 2,
  Relative_Figure_4_Parameter_Plot_Height = 2,
  Relative_Figure_4_Treatment_Plot_Dimension = 1,
  Relative_Figure_4_Individual_Plot_Height = 5,
  Relative_Figure_4_Individual_Plot_Width = 5,
  Relative_Figure_4_Parameter_Spacing_Line_Plot_Height = 1,
  Figure_4_Figure_Height_Constant = 75,
  Figure_4_Figure_Width_Constant = 100,
  Figure_4_Axis_Labels_and_Axis_Titles_Shifting_Constant = 1.375,
  Figure_4_Parameter_Title_Text_Size = 4,
  Figure_4_Treatment_Title_Text_Size = 3,
  Figure_4_Axis_Titles_Text_Size = 2.5,
  Figure_4_Axis_Labels_Text_Size = 2.5,
  Figure_4_Individual_Plot_Bottom_Margin = 10,
  Figure_4_Individual_Plot_Left_Margin = 10,
  Figure_4_Individual_Plot_Top_Margin = 1,
  Figure_4_Individual_Plot_Right_Margin = 1,
  Figure_4_Moment_Text_Size = 1.75,
  Figure_4_Panel_Separation_Line_Width = 2,
  Tolerance = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{Treatment}{w}

\item{Replicate}{w}

\item{Predictor_Variable}{w}

\item{Predictor_Variable_Bin_Lower_Limits}{w}

\item{Predictor_Variable_Bin_Upper_Limits}{w}

\item{Response_Variable}{w}

\item{Beta_Distribution_Alpha_Parameter_Values}{w}

\item{Beta_Distribution_Beta_Parameter_Values}{w}

\item{Data_Frame}{w}

\item{Parameters}{w}

\item{Formula}{w}

\item{Color_Data}{w}

\item{Color_Opacity}{w}

\item{Number_of_Simulations_per_Experimental_Unit}{w}

\item{Number_of_Histogram_Bins}{w}

\item{Moments_to_Report}{w}

\item{Make_Figure_1}{w}

\item{Make_Figure_2}{w}

\item{Make_Figure_3}{w}

\item{Make_Figure_4}{w}

\item{Working_Directory}{w}

\item{First_Figure_Name}{w}

\item{Second_Figure_Name}{w}

\item{Third_Figure_Name}{w}

\item{Fourth_Figure_Name}{w}

\item{Show_Moments_in_Figure_1}{w}

\item{Show_Overlapping_Areas_in_Figure_2}{w}

\item{Show_Moments_in_Figure_3}{w}

\item{Show_Moments_in_Figure_4}{w}

\item{Moment_Rounding_Constant}{w}

\item{Overlapping_Area_Rounding_Constant}{w}

\item{Number_of_Bins_in_the_Figure_4_Histograms}{w}

\item{Maximum_Number_of_Iterations}{w}

\item{Relative_Figure_1_Parameter_Plot_Height}{w}

\item{Relative_Figure_1_Treatment_Plot_Height}{w}

\item{Relative_Figure_1_Replicate_Plot_Height}{w}

\item{Relative_Figure_1_Individual_Plot_Height}{w}

\item{Relative_Figure_1_Legend_Plot_Height}{w}

\item{Relative_Figure_1_Individual_Plot_Width}{w}

\item{Relative_Figure_1_Gap_Plot_Width}{w}

\item{Relative_Figure_1_Parameter_Spacing_Line_Plot_Height}{w}

\item{Figure_1_Figure_Height_Constant}{w}

\item{Figure_1_Figure_Width_Constant}{w}

\item{Figure_1_Axis_Labels_and_Axis_Titles_Shifting_Constant}{w}

\item{Figure_1_Parameter_Title_Text_Size}{w}

\item{Figure_1_Treatment_Title_Text_Size}{w}

\item{Figure_1_Replicate_Title_Text_Size}{w}

\item{Figure_1_Axis_Titles_Text_Size}{w}

\item{Figure_1_Axis_Labels_Text_Size}{w}

\item{Figure_1_Individual_Plot_Bottom_Margin}{w}

\item{Figure_1_Individual_Plot_Left_Margin}{w}

\item{Figure_1_Individual_Plot_Top_Margin}{w}

\item{Figure_1_Individual_Plot_Right_Margin}{w}

\item{Figure_1_Moment_Text_Size}{w}

\item{Figure_1_Legend_Text_Size}{w}

\item{Figure_1_Panel_Separation_Line_Width}{w}

\item{Relative_Figure_2_Parameter_Plot_Height}{w}

\item{Relative_Figure_2_Treatment_Plot_Height}{w}

\item{Relative_Figure_2_Replicate_Plot_Height}{w}

\item{Relative_Figure_2_Individual_Plot_Dimension}{w}

\item{Relative_Figure_2_Gap_Plot_Dimension}{w}

\item{Relative_Figure_2_Legend_Plot_Height}{w}

\item{Relative_Figure_2_Parameter_Spacing_Line_Plot_Height}{w}

\item{Figure_2_Figure_Height_Constant}{w}

\item{Figure_2_Figure_Width_Constant}{w}

\item{Figure_2_Axis_Labels_and_Axis_Titles_Shifting_Constant}{w}

\item{Figure_2_Parameter_Title_Text_Size}{w}

\item{Figure_2_Treatment_Title_Text_Size}{w}

\item{Figure_2_Replicate_Title_Text_Size}{w}

\item{Figure_2_Axis_Titles_Text_Size}{w}

\item{Figure_2_Axis_Labels_Text_Size}{w}

\item{Figure_2_Individual_Plot_Bottom_Margin}{w}

\item{Figure_2_Individual_Plot_Left_Margin}{w}

\item{Figure_2_Individual_Plot_Top_Margin}{w}

\item{Figure_2_Individual_Plot_Right_Margin}{w}

\item{Figure_2_Overlapping_Area_Text_Size}{w}

\item{Figure_2_Legend_Text_Size}{w}

\item{Figure_2_Panel_Separation_Line_Width}{w}

\item{Relative_Figure_3_Parameter_Plot_Height}{w}

\item{Relative_Figure_3_Treatment_Plot_Dimension}{w}

\item{Relative_Figure_3_Replicate_Plot_Dimension}{w}

\item{Relative_Figure_3_Individual_Plot_Dimension}{w}

\item{Relative_Figure_3_Gap_Plot_Dimension}{w}

\item{Relative_Figure_3_Legend_Plot_Height}{w}

\item{Relative_Figure_3_Parameter_Spacing_Line_Plot_Height}{w}

\item{Figure_3_Figure_Height_Constant}{w}

\item{Figure_3_Figure_Width_Constant}{w}

\item{Figure_3_Axis_Labels_and_Axis_Titles_Shifting_Constant}{w}

\item{Figure_3_Parameter_Title_Text_Size}{w}

\item{Figure_3_Treatment_Title_Text_Size}{w}

\item{Figure_3_Replicate_Title_Text_Size}{w}

\item{Figure_3_Axis_Titles_Text_Size}{w}

\item{Figure_3_Axis_Labels_Text_Size}{w}

\item{Figure_3_Individual_Plot_Bottom_Margin}{w}

\item{Figure_3_Individual_Plot_Left_Margin}{w}

\item{Figure_3_Individual_Plot_Top_Margin}{w}

\item{Figure_3_Individual_Plot_Right_Margin}{w}

\item{Figure_3_Moment_Text_Size}{w}

\item{Figure_3_Legend_Text_Size}{w}

\item{Figure_3_Panel_Separation_Line_Width}{w}

\item{Relative_Figure_4_Parameter_Plot_Height}{w}

\item{Relative_Figure_4_Treatment_Plot_Dimension}{w}

\item{Relative_Figure_4_Individual_Plot_Height}{w}

\item{Relative_Figure_4_Individual_Plot_Width}{w}

\item{Relative_Figure_4_Parameter_Spacing_Line_Plot_Height}{w}

\item{Figure_4_Figure_Height_Constant}{w}

\item{Figure_4_Figure_Width_Constant}{w}

\item{Figure_4_Axis_Labels_and_Axis_Titles_Shifting_Constant}{w}

\item{Figure_4_Parameter_Title_Text_Size}{w}

\item{Figure_4_Treatment_Title_Text_Size}{w}

\item{Figure_4_Axis_Titles_Text_Size}{w}

\item{Figure_4_Axis_Labels_Text_Size}{w}

\item{Figure_4_Individual_Plot_Bottom_Margin}{w}

\item{Figure_4_Individual_Plot_Left_Margin}{w}

\item{Figure_4_Individual_Plot_Top_Margin}{w}

\item{Figure_4_Individual_Plot_Right_Margin}{w}

\item{Figure_4_Moment_Text_Size}{w}

\item{Figure_4_Panel_Separation_Line_Width}{w}
}
\value{
This function returns w
}
\description{
This function w
}
\details{
\code{Regression_for_Repeated_Measures_Designs_When_Predictor_Variables_Are_Bins_of_Numbers_and_Bins_Are_Continuous} w
}
\examples{
Data_Frame <- data.frame(Treatment = rep(c("Control", "Manure", "Synthetic Fertilizer"), each = 9), Block = rep(1:3, each = 3, length.out = 27), Soil_Depth = rep(c("0-5", "5-10", "10-20"), 9), Nitrogen_Concentration = c(5, 3, 2, 6, 6, 2, 5, 4, 1, 7, 4, 3, 6, 4, 3, 6, 5, 5, 8, 4, 4, 8, 5, 3, 7, 2, 1))
Data_Frame$Minimum_Depth <- as.numeric(gsub("-[[:digit:]]*$", "", Data_Frame$Soil_Depth))
Data_Frame$Maximum_Depth <- as.numeric(gsub("^[[:digit:]]*-", "", Data_Frame$Soil_Depth))
Data_Frame <- Data_Frame[, c(which(colnames(Data_Frame) == "Treatment"), which(colnames(Data_Frame) == "Block"), which(colnames(Data_Frame) == "Soil_Depth"), which(colnames(Data_Frame) == "Minimum_Depth"), which(colnames(Data_Frame) == "Maximum_Depth"), which(colnames(Data_Frame) == "Nitrogen_Concentration"))]
Color_Data <- data.frame(Treatment = c("Control", "Manure", "Synthetic Fertilizer"), Color = c("black", "red", "green"))
Linear_Regression_Parameters <- c("Slope", "Intercept")
Linear_Regression_Formula <- Nitrogen_Concentration ~ (Slope * Soil_Depth) + Intercept
Exponential_Regression_Parameters <- c("a", "b")
Exponential_Regression_Formula <- Nitrogen_Concentration ~ a * (b ^ Soil_Depth)
(Linear_Regression_Example <- Regression_for_Repeated_Measures_Designs_When_Predictor_Variables_Are_Bins_of_Numbers_and_Bins_Are_Continuous(Treatment = "Treatment", Replicate = "Block", Predictor_Variable_Bin_Lower_Limits = "Minimum_Depth", Predictor_Variable_Bin_Upper_Limits = "Maximum_Depth", Predictor_Variable = "Soil_Depth", Response_Variable = "Nitrogen_Concentration", Data_Frame = Data_Frame, Parameters = Linear_Regression_Parameters, Formula = Linear_Regression_Formula, Color_Data = Color_Data))
(Exponential_Regression_Example <- Regression_for_Repeated_Measures_Designs_When_Predictor_Variables_Are_Bins_of_Numbers_and_Bins_Are_Continuous(Treatment = "Treatment", Replicate = "Block", Predictor_Variable_Bin_Lower_Limits = "Minimum_Depth", Predictor_Variable_Bin_Upper_Limits = "Maximum_Depth", Predictor_Variable = "Soil_Depth", Response_Variable = "Nitrogen_Concentration", Data_Frame = Data_Frame, Parameters = Exponential_Regression_Parameters, Formula = Exponential_Regression_Formula, Color_Data = Color_Data))

}
\author{
David B. Moore (\email{davidblakneymoore@gmail.com})
}
