% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Removing_Matrix_Rows_and_Columns_Optimally.R
\name{Removing_Matrix_Rows_and_Columns_Optimally}
\alias{Removing_Matrix_Rows_and_Columns_Optimally}
\title{Removing Matrix Rows and Columns Optimally}
\usage{
Removing_Matrix_Rows_and_Columns_Optimally(
  Matrix,
  Values_to_Remove = "Missing Values"
)
}
\arguments{
\item{Matrix}{a matrix from which missing or non-finite values are to be removed.}

\item{Values_to_Remove}{one of either "Missing Values", "Non-Finite Values", or "Both Missing and Non-Finite Values"; indicates which types of values should be removed. For non-numeric matrices, only "Missing Values" may be provided.}
}
\value{
This function returns a matrix whose missing values or non-finite values have been removed by means of row and column deletion.
}
\description{
This function removes missing or non-finite values from matrices by means of row and column deletions. These deletions are performed in a way that ensures that the maximum number of good (non-missing or finite) values are retained in the resulting matrix.
}
\details{
\code{Removing_Matrix_Rows_and_Columns_Optimally} removes missing or non-finite values from matrices by means of row and column deletions. These deletions are performed in a way that ensures that the maximum number of good (non-missing or finite) values are retained in the resulting matrix.
}
\examples{
# Example 1 - a numeric matrix with both missing and non-finite values
set.seed(412)
Matrix_1 <- matrix(1:216, nrow = 12)
rownames(Matrix_1) <- paste(LETTERS[seq_len(nrow(Matrix_1))])
colnames(Matrix_1) <- paste(letters[seq_len(ncol(Matrix_1))])
Proportion_of_Missing_Values <- 0.2
for (i in seq_len(round(Proportion_of_Missing_Values * nrow(Matrix_1) * ncol(Matrix_1)))) {
  Matrix_1[sample(seq_len(nrow(Matrix_1)), 1), sample(seq_len(ncol(Matrix_1)), 1)] <- sample(c(NA, Inf, -Inf), 1, T, c(0.5, 0.25, 0.25))
}
Matrix_1
(Output_1 <- Removing_Matrix_Rows_and_Columns_Optimally(Matrix_1))

# Example 2 - a matrix of character strings with missing values
set.seed(410)
Matrix_2 <- matrix(LETTERS[1:24], ncol = 3)
for (i in seq_len(round(Proportion_of_Missing_Values * nrow(Matrix_2) * ncol(Matrix_2)))) {
  Matrix_2[sample(seq_len(nrow(Matrix_2)), 1), sample(seq_len(ncol(Matrix_2)), 1)] <- NA
}
rownames(Matrix_2) <- paste("Row", seq_len(nrow(Matrix_2)), sep = "_")
colnames(Matrix_2) <- paste("Column", seq_len(ncol(Matrix_2)), sep = "_")
Matrix_2
(Output_2 <- Removing_Matrix_Rows_and_Columns_Optimally(Matrix_2))

}
\author{
David B. Moore (\email{davidblakneymoore@gmail.com})
}
