% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Using_an_Improved_Bonferroni_Correction_for_Multiple_Comparisons.R
\name{Using_an_Improved_Bonferroni_Correction_for_Multiple_Comparisons}
\alias{Using_an_Improved_Bonferroni_Correction_for_Multiple_Comparisons}
\title{Using an Improved Bonferroni Correction for Multiple Comparisons}
\usage{
Using_an_Improved_Bonferroni_Correction_for_Multiple_Comparisons(
  Comparisons,
  p_Values,
  Data_Frame,
  Experimentwise_Type_I_Error_Rate = 0.05
)
}
\arguments{
\item{Comparisons}{w}

\item{p_Values}{w}

\item{Data_Frame}{w}

\item{Experimentwise_Type_I_Error_Rate}{w}
}
\value{
This function returns a data frame that contains the name of the comparisons or analyses, the original p values, the corrected p values (following this improved Bonferroni correction method), and a column indicating whether or not the corrected p values are statistically significant.
}
\description{
This function w
}
\details{
\code{Using_an_Improved_Bonferroni_Correction_for_Multiple_Comparisons} w
}
\examples{
# Generating Made-Up Data
Number_of_Comparisons <- 100
Data_Frame_1 <- data.frame(Comparison = paste('Comparison', seq_len(Number_of_Comparisons)), p_Value = rbeta(Number_of_Comparisons, 0.25, 1))

# Using the Function
Using_an_Improved_Bonferroni_Correction_for_Multiple_Comparisons(Comparison, p_Value, Data_Frame_1)

}
\author{
David B. Moore (\email{davidblakneymoore@gmail.com})
}
