% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aligning_Values_Across_Multiple_Vertical_Axes.R
\name{Aligning_Values_Across_Multiple_Vertical_Axes}
\alias{Aligning_Values_Across_Multiple_Vertical_Axes}
\title{Aligning Values Across Multiple Vertical Axes}
\usage{
Aligning_Values_Across_Multiple_Vertical_Axes(
  ...,
  Data_Frame,
  Values_to_Align = rep(0, length(list(...))),
  Variable_Weights = rep((1/length(list(...))), length(list(...))),
  Upper_Axis_Buffers = rep(0.05, length(list(...))),
  Lower_Axis_Buffers = rep(0.05, length(list(...)))
)
}
\arguments{
\item{...}{the numeric variables comprising the vertical axes.}

\item{Data_Frame}{an optional argument to provide if all the \code{...} arguments come from the same data frame and to prevent typing the data frame name multiple times.}

\item{Values_to_Align}{the values you wish to align across the vertical axes. The default for this argument is a vector of \code{0}s. This argument must be a numeric vector and it must contain the same number of elements that there are variables being aligned.}

\item{Variable_Weights}{the weights assigned to each variable. To prevent certain variables from being crowded near the top or the bottom of the plot, a greater weight can be assigned to these variables, which ensures that these variables will take up more of the plotting region (at the expense of other variables, of course). The default for this argument is to assign all the variables the same weight. This argument must be a numeric vector, all the entries must be finite and nonnegative, and it must contain the same number of elements as there are variables being aligned.}

\item{Upper_Axis_Buffers}{the minimum fractions of blank space you wish to leave around the top of the graph for each variable (above each variable's plotted points). The default for each variable is \code{0.05} - in other words, at least 5 \% of the space on the top of the graph will be empty above each variable's plotted points. This argument must be a numeric vector, all the entries must be between \code{0} and \code{1} (inclusive), and it must contain the same number of elements as there are variables being aligned.}

\item{Lower_Axis_Buffers}{the minimum fractions of blank space you wish to leave around the bottom of the graph for each variable (below each variable's plotted points). The default for each variable is \code{0.05} - in other words, at least 5 \% of the space on the bottom of the graph will be empty below each variable's plotted points. This argument must be a numeric vector, all the entries must be between \code{0} and \code{1} (inclusive), and it must contain the same number of elements as there are variables being aligned.}
}
\value{
This function returns the new axis limits for all the vertical axes being aligned.
}
\description{
This function generates new axis limits for multiple vertical axes such that specified values on the various vertical axes are aligned vertically. The user defines the values to align. Though the example at the end of the R script uses \code{base} R plotting functions, these new axis limits can be used with plotting functions from any package.
}
\details{
\code{Aligning_Values_Across_Multiple_Vertical_Axes} determines how to align values optimally across multiple vertical axes such that there isn't unnecessary  additional space on the top or the bottom of any graph and such that the specified values will align perfectly across all vertical axes.
}
\examples{
# Generate some made-up data.
set.seed(100)
Number_of_Rows <- 100
Index <- seq_len(Number_of_Rows)
Variable_1 <- rnorm(Number_of_Rows, -10, 1)
Variable_2 <- rnorm(Number_of_Rows, 0, 1)
Variable_3 <- rnorm(Number_of_Rows, 10, 1)
Made_Up_Data_for_the_Aligning_Values_Across_Multiple_Vertical_Axes_Function <- data.frame(Index = Index, Variable_1 = Variable_1, Variable_2 = Variable_2, Variable_3 = Variable_3)

# These are the values that will align on our graph.
Values_to_Align = c(-2, 0, 0)

# Weigh the first variable more heavily than the other two.
Variable_Weights <- c(0.75, 0.125, 0.125)

# Generate the new axis limits.
(Final_Vertical_Axis_Limits <- Aligning_Values_Across_Multiple_Vertical_Axes(Variable_1, Variable_2, Variable_3, Data_Frame = Made_Up_Data_for_the_Aligning_Values_Across_Multiple_Vertical_Axes_Function, Values_to_Align = Values_to_Align, Variable_Weights = Variable_Weights))

# Make some plots.
tryCatch (dev.off(), error = function (e) {
.
})
layout(matrix(c(1, 1, 2, 2, 3, 3, 4, 4, 4, 5, 5, 5, 6, 6, 6, 6, 6, 6), ncol = 6, byrow = TRUE), heights = c((3 / 9), (5 / 9), (1 / 9)))
par(mar = c(6, 5, 5, 3))
plot(Variable_1 ~ Index, data = Made_Up_Data_for_the_Aligning_Values_Across_Multiple_Vertical_Axes_Function, xlab = "", ylab = "", main = "Variable 1", pch = 20)
mtext("Index", 1, line = 2.5)
mtext("Variable 1", 2, line = 2.5)
plot(Variable_2 ~ Index, data = Made_Up_Data_for_the_Aligning_Values_Across_Multiple_Vertical_Axes_Function, xlab = "", ylab = "", main = "Variable 2", pch = 20, col = 2)
mtext("Index", 1, line = 2.5)
mtext("Variable 2", 2, line = 2.5)
plot(Variable_3 ~ Index, data = Made_Up_Data_for_the_Aligning_Values_Across_Multiple_Vertical_Axes_Function, xlab = "", ylab = "", main = "Variable 3", pch = 20, col = 3)
mtext("Index", 1, line = 2.5)
mtext("Variable 3", 2, line = 2.5)
par(mar = c(6, 6, 6, 12))
plot(Variable_1 ~ Index, data = Made_Up_Data_for_the_Aligning_Values_Across_Multiple_Vertical_Axes_Function, xlab = "", ylab = "", yaxt = "n", pch = 20, main = "All Variables on One Plot (Values not Aligned)")
axis(2, at = pretty(Made_Up_Data_for_the_Aligning_Values_Across_Multiple_Vertical_Axes_Function$Variable_1))
mtext("Index", 1, line = 2.5)
mtext("Variable 1", 2, line = 2.5)
par(new = T)
plot(Variable_2 ~ Index, data = Made_Up_Data_for_the_Aligning_Values_Across_Multiple_Vertical_Axes_Function, xlab = "", ylab = "", yaxt = "n", col = 2, pch = 20)
axis(4, at = pretty(Made_Up_Data_for_the_Aligning_Values_Across_Multiple_Vertical_Axes_Function$Variable_2))
mtext("Variable 2", 4, line = 2.5)
par(new = T)
plot(Variable_3 ~ Index, data = Made_Up_Data_for_the_Aligning_Values_Across_Multiple_Vertical_Axes_Function, xlab = "", ylab = "", yaxt = "n", col = 3, pch = 20)
axis(4, at = pretty(Made_Up_Data_for_the_Aligning_Values_Across_Multiple_Vertical_Axes_Function$Variable_3), line = 5)
mtext("Variable 3", 4, line = 7.5)
plot(Variable_1 ~ Index, data = Made_Up_Data_for_the_Aligning_Values_Across_Multiple_Vertical_Axes_Function, xlab = "", ylab = "", yaxt = "n", ylim = Final_Vertical_Axis_Limits[[1]], pch = 20, main = "All Variables on One Plot (Values Aligned)")
axis(2, at = pretty(range(Final_Vertical_Axis_Limits[[1]])))
mtext("Index", 1, line = 2.5)
mtext("Variable 1", 2, line = 2.5)
abline(h = Values_to_Align[1], lty = 2)
par(new = T)
plot(Variable_2 ~ Index, data = Made_Up_Data_for_the_Aligning_Values_Across_Multiple_Vertical_Axes_Function, xlab = "", ylab = "", yaxt = "n", ylim = Final_Vertical_Axis_Limits[[2]], col = 2, pch = 20)
axis(4, at = pretty(range(Final_Vertical_Axis_Limits[[2]])))
mtext("Variable 2", 4, line = 2.5)
abline(h = Values_to_Align[2], lty = 2)
par(new = T)
plot(Variable_3 ~ Index, data = Made_Up_Data_for_the_Aligning_Values_Across_Multiple_Vertical_Axes_Function, xlab = "", ylab = "", yaxt = "n", ylim = Final_Vertical_Axis_Limits[[3]], col = 3, pch = 20)
axis(4, at = pretty(range(Final_Vertical_Axis_Limits[[3]])), line = 5)
mtext("Variable 3", 4, line = 7.5)
abline(h = Values_to_Align[3], lty = 2)
par(mar = c(0, 0, 0, 0))
plot(0, type = "n", axes = FALSE, xlab = "", ylab = "")
legend("center", xpd = TRUE, inset = c(0, -0.3), title = expression(paste(bold("Variable"))), legend = 1:3, col = 1:3, pch = 20, horiz = T)

}
\author{
David B. Moore (\email{davidblakneymoore@gmail.com})
}
